% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{ActionSetting}
\alias{ActionSetting}
\title{Mouse action settings on a shape}
\description{
Provides access to click and hover action properties of a shape, including
hyperlink URL and action type. Access via \code{shape$click_action}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActionSetting-new}{\code{ActionSetting$new()}}
\item \href{#method-ActionSetting-clone}{\code{ActionSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActionSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActionSetting-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionSetting$new(cNvPr, parent, hover = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActionSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActionSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
