% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_inches}
\alias{as_inches}
\alias{as_cm}
\alias{as_mm}
\alias{as_pt}
\alias{as_emu}
\alias{as_centipoints}
\title{Convert a Length value to inches}
\usage{
as_inches(x)

as_cm(x)

as_mm(x)

as_pt(x)

as_emu(x)

as_centipoints(x)
}
\arguments{
\item{x}{A Length value (integer EMU).}
}
\value{
Numeric value in the target unit.
}
\description{
Convert a Length value to inches
}
