% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{AxisTitle}
\alias{AxisTitle}
\title{Axis title domain object}
\description{
Axis title domain object

Axis title domain object
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AxisTitle-new}{\code{AxisTitle$new()}}
\item \href{#method-AxisTitle-clone}{\code{AxisTitle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AxisTitle-new"></a>}}
\if{latex}{\out{\hypertarget{method-AxisTitle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxisTitle$new(title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AxisTitle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AxisTitle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxisTitle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
