% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{BaseAxis}
\alias{BaseAxis}
\title{Base class for chart axis domain objects}
\description{
Base class for chart axis domain objects

Base class for chart axis domain objects
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseAxis-new}{\code{BaseAxis$new()}}
\item \href{#method-BaseAxis-clone}{\code{BaseAxis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxis-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxis$new(xAx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
