% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{BaseAxisElement}
\alias{BaseAxisElement}
\title{Base class for chart axis XML elements}
\description{
Base class for chart axis XML elements

Base class for chart axis XML elements
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{BaseAxisElement}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseAxisElement-get_or_add_txPr}{\code{BaseAxisElement$get_or_add_txPr()}}
\item \href{#method-BaseAxisElement-get_or_add_title}{\code{BaseAxisElement$get_or_add_title()}}
\item \href{#method-BaseAxisElement-.remove_title}{\code{BaseAxisElement$.remove_title()}}
\item \href{#method-BaseAxisElement-get_or_add_majorGridlines}{\code{BaseAxisElement$get_or_add_majorGridlines()}}
\item \href{#method-BaseAxisElement-.remove_majorGridlines}{\code{BaseAxisElement$.remove_majorGridlines()}}
\item \href{#method-BaseAxisElement-get_or_add_minorGridlines}{\code{BaseAxisElement$get_or_add_minorGridlines()}}
\item \href{#method-BaseAxisElement-.remove_minorGridlines}{\code{BaseAxisElement$.remove_minorGridlines()}}
\item \href{#method-BaseAxisElement-get_or_add_tickLblPos}{\code{BaseAxisElement$get_or_add_tickLblPos()}}
\item \href{#method-BaseAxisElement-get_or_add_numFmt}{\code{BaseAxisElement$get_or_add_numFmt()}}
\item \href{#method-BaseAxisElement-.add_majorTickMark}{\code{BaseAxisElement$.add_majorTickMark()}}
\item \href{#method-BaseAxisElement-.remove_majorTickMark}{\code{BaseAxisElement$.remove_majorTickMark()}}
\item \href{#method-BaseAxisElement-.add_minorTickMark}{\code{BaseAxisElement$.add_minorTickMark()}}
\item \href{#method-BaseAxisElement-.remove_minorTickMark}{\code{BaseAxisElement$.remove_minorTickMark()}}
\item \href{#method-BaseAxisElement-get_or_add_scaling}{\code{BaseAxisElement$get_or_add_scaling()}}
\item \href{#method-BaseAxisElement-get_or_add_delete_}{\code{BaseAxisElement$get_or_add_delete_()}}
\item \href{#method-BaseAxisElement-get_or_add_spPr}{\code{BaseAxisElement$get_or_add_spPr()}}
\item \href{#method-BaseAxisElement-clone}{\code{BaseAxisElement$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_txPr"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_txPr}{}}}
\subsection{Method \code{get_or_add_txPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_txPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_title"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_title}{}}}
\subsection{Method \code{get_or_add_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.remove_title"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.remove_title}{}}}
\subsection{Method \code{.remove_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.remove_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_majorGridlines"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_majorGridlines}{}}}
\subsection{Method \code{get_or_add_majorGridlines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_majorGridlines()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.remove_majorGridlines"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.remove_majorGridlines}{}}}
\subsection{Method \code{.remove_majorGridlines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.remove_majorGridlines()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_minorGridlines"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_minorGridlines}{}}}
\subsection{Method \code{get_or_add_minorGridlines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_minorGridlines()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.remove_minorGridlines"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.remove_minorGridlines}{}}}
\subsection{Method \code{.remove_minorGridlines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.remove_minorGridlines()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_tickLblPos"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_tickLblPos}{}}}
\subsection{Method \code{get_or_add_tickLblPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_tickLblPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_numFmt"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_numFmt}{}}}
\subsection{Method \code{get_or_add_numFmt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_numFmt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.add_majorTickMark"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.add_majorTickMark}{}}}
\subsection{Method \code{.add_majorTickMark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.add_majorTickMark(val = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.remove_majorTickMark"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.remove_majorTickMark}{}}}
\subsection{Method \code{.remove_majorTickMark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.remove_majorTickMark()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.add_minorTickMark"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.add_minorTickMark}{}}}
\subsection{Method \code{.add_minorTickMark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.add_minorTickMark(val = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-.remove_minorTickMark"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-.remove_minorTickMark}{}}}
\subsection{Method \code{.remove_minorTickMark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$.remove_minorTickMark()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_scaling"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_scaling}{}}}
\subsection{Method \code{get_or_add_scaling()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_scaling()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_delete_"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_delete_}{}}}
\subsection{Method \code{get_or_add_delete_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_delete_()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-get_or_add_spPr"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-get_or_add_spPr}{}}}
\subsection{Method \code{get_or_add_spPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$get_or_add_spPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAxisElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAxisElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAxisElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
