% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{BaseChartData}
\alias{BaseChartData}
\title{Base class for chart data objects}
\description{
Base class for chart data objects

Base class for chart data objects
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseChartData-new}{\code{BaseChartData$new()}}
\item \href{#method-BaseChartData-get}{\code{BaseChartData$get()}}
\item \href{#method-BaseChartData-to_list}{\code{BaseChartData$to_list()}}
\item \href{#method-BaseChartData-data_point_offset}{\code{BaseChartData$data_point_offset()}}
\item \href{#method-BaseChartData-series_index}{\code{BaseChartData$series_index()}}
\item \href{#method-BaseChartData-series_name_ref}{\code{BaseChartData$series_name_ref()}}
\item \href{#method-BaseChartData-x_values_ref}{\code{BaseChartData$x_values_ref()}}
\item \href{#method-BaseChartData-y_values_ref}{\code{BaseChartData$y_values_ref()}}
\item \href{#method-BaseChartData-xml_str}{\code{BaseChartData$xml_str()}}
\item \href{#method-BaseChartData-.workbook_writer}{\code{BaseChartData$.workbook_writer()}}
\item \href{#method-BaseChartData-clone}{\code{BaseChartData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$new(number_format = "General")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-get"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$get(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-data_point_offset"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-data_point_offset}{}}}
\subsection{Method \code{data_point_offset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$data_point_offset(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-series_index"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-series_index}{}}}
\subsection{Method \code{series_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$series_index(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-series_name_ref"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-series_name_ref}{}}}
\subsection{Method \code{series_name_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$series_name_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-x_values_ref"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-x_values_ref}{}}}
\subsection{Method \code{x_values_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$x_values_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-y_values_ref"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-y_values_ref}{}}}
\subsection{Method \code{y_values_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$y_values_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-xml_str"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-xml_str}{}}}
\subsection{Method \code{xml_str()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$xml_str(chart_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-.workbook_writer"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-.workbook_writer}{}}}
\subsection{Method \code{.workbook_writer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$.workbook_writer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseChartData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseChartData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseChartData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
