% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-xmlchemy.R
\name{BaseOxmlElement}
\alias{BaseOxmlElement}
\title{Base class for all custom XML element classes}
\description{
Wraps an xml2 xml_node and provides standardized methods for child element
access, insertion, removal, and attribute manipulation.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseOxmlElement-new}{\code{BaseOxmlElement$new()}}
\item \href{#method-BaseOxmlElement-get_node}{\code{BaseOxmlElement$get_node()}}
\item \href{#method-BaseOxmlElement-find}{\code{BaseOxmlElement$find()}}
\item \href{#method-BaseOxmlElement-findall}{\code{BaseOxmlElement$findall()}}
\item \href{#method-BaseOxmlElement-get_attr}{\code{BaseOxmlElement$get_attr()}}
\item \href{#method-BaseOxmlElement-set_attr}{\code{BaseOxmlElement$set_attr()}}
\item \href{#method-BaseOxmlElement-remove_attr}{\code{BaseOxmlElement$remove_attr()}}
\item \href{#method-BaseOxmlElement-get_attrs}{\code{BaseOxmlElement$get_attrs()}}
\item \href{#method-BaseOxmlElement-append_child}{\code{BaseOxmlElement$append_child()}}
\item \href{#method-BaseOxmlElement-remove_child}{\code{BaseOxmlElement$remove_child()}}
\item \href{#method-BaseOxmlElement-first_child_found_in}{\code{BaseOxmlElement$first_child_found_in()}}
\item \href{#method-BaseOxmlElement-insert_element_before}{\code{BaseOxmlElement$insert_element_before()}}
\item \href{#method-BaseOxmlElement-remove_all}{\code{BaseOxmlElement$remove_all()}}
\item \href{#method-BaseOxmlElement-xpath}{\code{BaseOxmlElement$xpath()}}
\item \href{#method-BaseOxmlElement-to_xml}{\code{BaseOxmlElement$to_xml()}}
\item \href{#method-BaseOxmlElement-clone}{\code{BaseOxmlElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$new(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-get_node"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-get_node}{}}}
\subsection{Method \code{get_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$get_node()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-find"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-find}{}}}
\subsection{Method \code{find()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$find(clark_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-findall"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-findall}{}}}
\subsection{Method \code{findall()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$findall(clark_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-get_attr"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-get_attr}{}}}
\subsection{Method \code{get_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$get_attr(attr_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-set_attr"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-set_attr}{}}}
\subsection{Method \code{set_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$set_attr(attr_name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-remove_attr"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-remove_attr}{}}}
\subsection{Method \code{remove_attr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$remove_attr(attr_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$get_attrs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-append_child"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-append_child}{}}}
\subsection{Method \code{append_child()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$append_child(child)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-remove_child"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-remove_child}{}}}
\subsection{Method \code{remove_child()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$remove_child(child)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-first_child_found_in"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-first_child_found_in}{}}}
\subsection{Method \code{first_child_found_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$first_child_found_in(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-insert_element_before"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-insert_element_before}{}}}
\subsection{Method \code{insert_element_before()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$insert_element_before(elm, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-remove_all"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-remove_all}{}}}
\subsection{Method \code{remove_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$remove_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-xpath"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-xpath}{}}}
\subsection{Method \code{xpath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$xpath(xpath_str)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-to_xml"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-to_xml}{}}}
\subsection{Method \code{to_xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$to_xml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseOxmlElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseOxmlElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseOxmlElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
