% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{BasePlot}
\alias{BasePlot}
\title{Base class for chart plot objects}
\description{
Base class for chart plot objects

Base class for chart plot objects
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasePlot-new}{\code{BasePlot$new()}}
\item \href{#method-BasePlot-clone}{\code{BasePlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasePlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasePlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePlot$new(xChart, chart)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasePlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasePlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
