% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{BaseSeries}
\alias{BaseSeries}
\title{Base class for chart series objects}
\description{
Base class for chart series objects

Base class for chart series objects
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseSeries-new}{\code{BaseSeries$new()}}
\item \href{#method-BaseSeries-clone}{\code{BaseSeries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseSeries-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseSeries-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseSeries$new(ser)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseSeries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseSeries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
