% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{BaseSeriesData}
\alias{BaseSeriesData}
\title{Base class for series data objects}
\description{
Base class for series data objects

Base class for series data objects
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseSeriesData-new}{\code{BaseSeriesData$new()}}
\item \href{#method-BaseSeriesData-clone}{\code{BaseSeriesData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseSeriesData-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseSeriesData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseSeriesData$new(chart_data, name, number_format = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseSeriesData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseSeriesData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseSeriesData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
