% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{BaseShape}
\alias{BaseShape}
\title{Base class for shape proxy objects}
\description{
Subclasses include Shape, Picture, Connector, GroupShape, GraphicFrame.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseShape-new}{\code{BaseShape$new()}}
\item \href{#method-BaseShape-clone}{\code{BaseShape$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseShape-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseShape-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseShape$new(shape_elm, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseShape-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseShape-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseShape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
