% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{BubbleChartData}
\alias{BubbleChartData}
\title{Chart data container for bubble charts}
\description{
Holds series and bubble data points (x, y, bubble size). Use \code{add_series()}
to add a series, then call \code{add_data_point()} on each returned series object.
}
\section{Super classes}{
\code{rpptx::BaseChartData} -> \code{\link[rpptx:XyChartData]{rpptx::XyChartData}} -> \code{BubbleChartData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BubbleChartData-add_series}{\code{BubbleChartData$add_series()}}
\item \href{#method-BubbleChartData-bubble_sizes_ref}{\code{BubbleChartData$bubble_sizes_ref()}}
\item \href{#method-BubbleChartData-.workbook_writer}{\code{BubbleChartData$.workbook_writer()}}
\item \href{#method-BubbleChartData-clone}{\code{BubbleChartData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="data_point_offset"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-data_point_offset'><code>rpptx::BaseChartData$data_point_offset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="get"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-get'><code>rpptx::BaseChartData$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="initialize"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-initialize'><code>rpptx::BaseChartData$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="series_index"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-series_index'><code>rpptx::BaseChartData$series_index()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="series_name_ref"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-series_name_ref'><code>rpptx::BaseChartData$series_name_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="to_list"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-to_list'><code>rpptx::BaseChartData$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="x_values_ref"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-x_values_ref'><code>rpptx::BaseChartData$x_values_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="xml_str"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-xml_str'><code>rpptx::BaseChartData$xml_str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="y_values_ref"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-y_values_ref'><code>rpptx::BaseChartData$y_values_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleChartData-add_series"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleChartData-add_series}{}}}
\subsection{Method \code{add_series()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleChartData$add_series(name, number_format = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleChartData-bubble_sizes_ref"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleChartData-bubble_sizes_ref}{}}}
\subsection{Method \code{bubble_sizes_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleChartData$bubble_sizes_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleChartData-.workbook_writer"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleChartData-.workbook_writer}{}}}
\subsection{Method \code{.workbook_writer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleChartData$.workbook_writer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleChartData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleChartData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleChartData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
