% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{BubblePlot}
\alias{BubblePlot}
\title{Bubble chart plot}
\description{
Bubble chart plot

Bubble chart plot
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BasePlot]{rpptx::BasePlot}} -> \code{BubblePlot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BubblePlot-clone}{\code{BubblePlot$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BasePlot" data-id="initialize"><a href='../../rpptx/html/BasePlot.html#method-BasePlot-initialize'><code>rpptx::BasePlot$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubblePlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BubblePlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubblePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
