% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{BubblePoints}
\alias{BubblePoints}
\title{Point collection for bubble series}
\description{
Point collection for bubble series

Point collection for bubble series
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BubblePoints-new}{\code{BubblePoints$new()}}
\item \href{#method-BubblePoints-get}{\code{BubblePoints$get()}}
\item \href{#method-BubblePoints-clone}{\code{BubblePoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubblePoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-BubblePoints-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubblePoints$new(ser)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubblePoints-get"></a>}}
\if{latex}{\out{\hypertarget{method-BubblePoints-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubblePoints$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubblePoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BubblePoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubblePoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
