% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{BubbleSeriesData}
\alias{BubbleSeriesData}
\title{Series data for a bubble chart}
\description{
Series data for a bubble chart

Series data for a bubble chart
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:BaseSeriesData]{rpptx::BaseSeriesData}} -> \code{\link[rpptx:XySeriesData]{rpptx::XySeriesData}} -> \code{BubbleSeriesData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BubbleSeriesData-add_data_point}{\code{BubbleSeriesData$add_data_point()}}
\item \href{#method-BubbleSeriesData-clone}{\code{BubbleSeriesData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseSeriesData" data-id="initialize"><a href='../../rpptx/html/BaseSeriesData.html#method-BaseSeriesData-initialize'><code>rpptx::BaseSeriesData$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleSeriesData-add_data_point"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleSeriesData-add_data_point}{}}}
\subsection{Method \code{add_data_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleSeriesData$add_data_point(x, y, size, number_format = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleSeriesData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleSeriesData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleSeriesData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
