% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-xlsx.R
\name{BubbleWorkbookWriter}
\alias{BubbleWorkbookWriter}
\title{Workbook writer for bubble charts}
\description{
Workbook writer for bubble charts

Workbook writer for bubble charts
}
\keyword{internal}
\section{Super classes}{
\code{rpptx::.BaseWorkbookWriter} -> \code{\link[rpptx:XyWorkbookWriter]{rpptx::XyWorkbookWriter}} -> \code{BubbleWorkbookWriter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BubbleWorkbookWriter-bubble_sizes_ref}{\code{BubbleWorkbookWriter$bubble_sizes_ref()}}
\item \href{#method-BubbleWorkbookWriter-clone}{\code{BubbleWorkbookWriter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic=".BaseWorkbookWriter" data-id="initialize"><a href='../../rpptx/html/.BaseWorkbookWriter.html#method-.BaseWorkbookWriter-initialize'><code>rpptx::.BaseWorkbookWriter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XyWorkbookWriter" data-id="series_name_ref"><a href='../../rpptx/html/XyWorkbookWriter.html#method-XyWorkbookWriter-series_name_ref'><code>rpptx::XyWorkbookWriter$series_name_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XyWorkbookWriter" data-id="x_values_ref"><a href='../../rpptx/html/XyWorkbookWriter.html#method-XyWorkbookWriter-x_values_ref'><code>rpptx::XyWorkbookWriter$x_values_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XyWorkbookWriter" data-id="y_values_ref"><a href='../../rpptx/html/XyWorkbookWriter.html#method-XyWorkbookWriter-y_values_ref'><code>rpptx::XyWorkbookWriter$y_values_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleWorkbookWriter-bubble_sizes_ref"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleWorkbookWriter-bubble_sizes_ref}{}}}
\subsection{Method \code{bubble_sizes_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleWorkbookWriter$bubble_sizes_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BubbleWorkbookWriter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BubbleWorkbookWriter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BubbleWorkbookWriter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
