% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-serialized.R
\name{CaseInsensitiveDict}
\alias{CaseInsensitiveDict}
\title{A simple case-insensitive named list}
\usage{
CaseInsensitiveDict(...)
}
\arguments{
\item{...}{Key-value pairs.}
}
\value{
An environment with case-insensitive access.
}
\description{
A simple case-insensitive named list
}
\keyword{internal}
