% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{Categories}
\alias{Categories}
\title{Ordered collection of category labels for a chart}
\description{
Ordered collection of category labels for a chart

Ordered collection of category labels for a chart
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Categories-new}{\code{Categories$new()}}
\item \href{#method-Categories-add_category}{\code{Categories$add_category()}}
\item \href{#method-Categories-index}{\code{Categories$index()}}
\item \href{#method-Categories-get}{\code{Categories$get()}}
\item \href{#method-Categories-clone}{\code{Categories$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Categories-new"></a>}}
\if{latex}{\out{\hypertarget{method-Categories-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Categories$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Categories-add_category"></a>}}
\if{latex}{\out{\hypertarget{method-Categories-add_category}{}}}
\subsection{Method \code{add_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Categories$add_category(label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Categories-index"></a>}}
\if{latex}{\out{\hypertarget{method-Categories-index}{}}}
\subsection{Method \code{index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Categories$index(cat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Categories-get"></a>}}
\if{latex}{\out{\hypertarget{method-Categories-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Categories$get(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Categories-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Categories-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Categories$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
