% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{Category}
\alias{Category}
\title{A single chart category label}
\description{
A single chart category label

A single chart category label
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Category-new}{\code{Category$new()}}
\item \href{#method-Category-add_sub_category}{\code{Category$add_sub_category()}}
\item \href{#method-Category-index}{\code{Category$index()}}
\item \href{#method-Category-numeric_str_val}{\code{Category$numeric_str_val()}}
\item \href{#method-Category-clone}{\code{Category$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-new"></a>}}
\if{latex}{\out{\hypertarget{method-Category-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$new(label, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-add_sub_category"></a>}}
\if{latex}{\out{\hypertarget{method-Category-add_sub_category}{}}}
\subsection{Method \code{add_sub_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$add_sub_category(label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-index"></a>}}
\if{latex}{\out{\hypertarget{method-Category-index}{}}}
\subsection{Method \code{index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$index(sub_cat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-numeric_str_val"></a>}}
\if{latex}{\out{\hypertarget{method-Category-numeric_str_val}{}}}
\subsection{Method \code{numeric_str_val()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$numeric_str_val(date_1904 = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Category-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
