% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{CategoryPoints}
\alias{CategoryPoints}
\title{Point collection for category-based series}
\description{
Point collection for category-based series

Point collection for category-based series
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategoryPoints-new}{\code{CategoryPoints$new()}}
\item \href{#method-CategoryPoints-get}{\code{CategoryPoints$get()}}
\item \href{#method-CategoryPoints-clone}{\code{CategoryPoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryPoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryPoints-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryPoints$new(ser)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryPoints-get"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryPoints-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryPoints$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryPoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryPoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryPoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
