% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{CategorySeriesData}
\alias{CategorySeriesData}
\title{Series data for a category chart}
\description{
Series data for a category chart

Series data for a category chart
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseSeriesData]{rpptx::BaseSeriesData}} -> \code{CategorySeriesData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategorySeriesData-add_data_point}{\code{CategorySeriesData$add_data_point()}}
\item \href{#method-CategorySeriesData-clone}{\code{CategorySeriesData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseSeriesData" data-id="initialize"><a href='../../rpptx/html/BaseSeriesData.html#method-BaseSeriesData-initialize'><code>rpptx::BaseSeriesData$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategorySeriesData-add_data_point"></a>}}
\if{latex}{\out{\hypertarget{method-CategorySeriesData-add_data_point}{}}}
\subsection{Method \code{add_data_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategorySeriesData$add_data_point(value, number_format = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategorySeriesData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategorySeriesData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategorySeriesData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
