% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-xlsx.R
\name{CategoryWorkbookWriter}
\alias{CategoryWorkbookWriter}
\title{Workbook writer for category (bar, column, line, pie, etc.) charts}
\description{
Workbook writer for category (bar, column, line, pie, etc.) charts

Workbook writer for category (bar, column, line, pie, etc.) charts
}
\keyword{internal}
\section{Super class}{
\code{rpptx::.BaseWorkbookWriter} -> \code{CategoryWorkbookWriter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategoryWorkbookWriter-categories_ref}{\code{CategoryWorkbookWriter$categories_ref()}}
\item \href{#method-CategoryWorkbookWriter-series_name_ref}{\code{CategoryWorkbookWriter$series_name_ref()}}
\item \href{#method-CategoryWorkbookWriter-values_ref}{\code{CategoryWorkbookWriter$values_ref()}}
\item \href{#method-CategoryWorkbookWriter-clone}{\code{CategoryWorkbookWriter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic=".BaseWorkbookWriter" data-id="initialize"><a href='../../rpptx/html/.BaseWorkbookWriter.html#method-.BaseWorkbookWriter-initialize'><code>rpptx::.BaseWorkbookWriter$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryWorkbookWriter-categories_ref"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryWorkbookWriter-categories_ref}{}}}
\subsection{Method \code{categories_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryWorkbookWriter$categories_ref()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryWorkbookWriter-series_name_ref"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryWorkbookWriter-series_name_ref}{}}}
\subsection{Method \code{series_name_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryWorkbookWriter$series_name_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryWorkbookWriter-values_ref"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryWorkbookWriter-values_ref}{}}}
\subsection{Method \code{values_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryWorkbookWriter$values_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryWorkbookWriter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryWorkbookWriter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoryWorkbookWriter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
