% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{Chart}
\alias{Chart}
\title{Chart domain object}
\description{
Provides access to chart properties, title, legend, plots, series, and axes.
Access via \code{graphic_frame$chart}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Chart-new}{\code{Chart$new()}}
\item \href{#method-Chart-clone}{\code{Chart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chart-new"></a>}}
\if{latex}{\out{\hypertarget{method-Chart-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chart$new(chartSpace, chart_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Chart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
