% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-xmlwriter.R
\name{chart_xml_writer}
\alias{chart_xml_writer}
\title{Return the appropriate chart XML writer for chart_type}
\usage{
chart_xml_writer(chart_type, chart_data)
}
\arguments{
\item{chart_type}{An integer from XL_CHART_TYPE, e.g. \code{XL_CHART_TYPE$COLUMN_CLUSTERED}.}

\item{chart_data}{A \link{CategoryChartData}, \link{XyChartData}, or \link{BubbleChartData} object.}
}
\value{
An R6 writer object with an \verb{$xml} active binding returning the XML string.
}
\description{
Return the appropriate chart XML writer for chart_type
}
\keyword{internal}
