% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-data.R
\name{ChartData}
\alias{ChartData}
\title{Alias: ChartData == CategoryChartData}
\description{
Alias: ChartData == CategoryChartData

Alias: ChartData == CategoryChartData
}
\section{Super class}{
\code{\link[rpptx:BaseChartData]{rpptx::BaseChartData}} -> \code{CategoryChartData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategoryChartData-add_series}{\code{ChartData$add_series()}}
\item \href{#method-CategoryChartData-add_category}{\code{ChartData$add_category()}}
\item \href{#method-CategoryChartData-categories_ref}{\code{ChartData$categories_ref()}}
\item \href{#method-CategoryChartData-values_ref}{\code{ChartData$values_ref()}}
\item \href{#method-CategoryChartData-.workbook_writer}{\code{ChartData$.workbook_writer()}}
\item \href{#method-CategoryChartData-clone}{\code{ChartData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="data_point_offset"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-data_point_offset'><code>rpptx::BaseChartData$data_point_offset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="get"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-get'><code>rpptx::BaseChartData$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="initialize"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-initialize'><code>rpptx::BaseChartData$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="series_index"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-series_index'><code>rpptx::BaseChartData$series_index()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="series_name_ref"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-series_name_ref'><code>rpptx::BaseChartData$series_name_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="to_list"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-to_list'><code>rpptx::BaseChartData$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="x_values_ref"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-x_values_ref'><code>rpptx::BaseChartData$x_values_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="xml_str"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-xml_str'><code>rpptx::BaseChartData$xml_str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseChartData" data-id="y_values_ref"><a href='../../rpptx/html/BaseChartData.html#method-BaseChartData-y_values_ref'><code>rpptx::BaseChartData$y_values_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryChartData-add_series"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryChartData-add_series}{}}}
\subsection{Method \code{add_series()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartData$add_series(name, values = numeric(0), number_format = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryChartData-add_category"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryChartData-add_category}{}}}
\subsection{Method \code{add_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartData$add_category(label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryChartData-categories_ref"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryChartData-categories_ref}{}}}
\subsection{Method \code{categories_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartData$categories_ref()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryChartData-values_ref"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryChartData-values_ref}{}}}
\subsection{Method \code{values_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartData$values_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryChartData-.workbook_writer"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryChartData-.workbook_writer}{}}}
\subsection{Method \code{.workbook_writer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartData$.workbook_writer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoryChartData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategoryChartData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
