% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{ChartFormat}
\alias{ChartFormat}
\title{Chart element format (fill + line)}
\description{
Provides \code{fill} and \code{line} properties for chart sub-elements like axes,
series, markers, and gridlines.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChartFormat-new}{\code{ChartFormat$new()}}
\item \href{#method-ChartFormat-clone}{\code{ChartFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChartFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartFormat$new(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChartFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
