% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-chart.R
\name{ChartPart}
\alias{ChartPart}
\title{Chart XML part}
\description{
Wraps a c:chartSpace XML document. Created via \code{\link[=ChartPart_new]{ChartPart_new()}}.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:Part]{rpptx::Part}} -> \code{\link[rpptx:XmlPart]{rpptx::XmlPart}} -> \code{ChartPart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChartPart-load}{\code{ChartPart$load()}}
\item \href{#method-ChartPart-clone}{\code{ChartPart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="has_rels"><a href='../../rpptx/html/Part.html#method-Part-has_rels'><code>rpptx::Part$has_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="load_rels_from_xml"><a href='../../rpptx/html/Part.html#method-Part-load_rels_from_xml'><code>rpptx::Part$load_rels_from_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="part_related_by"><a href='../../rpptx/html/Part.html#method-Part-part_related_by'><code>rpptx::Part$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="relate_to"><a href='../../rpptx/html/Part.html#method-Part-relate_to'><code>rpptx::Part$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="related_part"><a href='../../rpptx/html/Part.html#method-Part-related_part'><code>rpptx::Part$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="rels_xml_bytes"><a href='../../rpptx/html/Part.html#method-Part-rels_xml_bytes'><code>rpptx::Part$rels_xml_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="target_ref"><a href='../../rpptx/html/Part.html#method-Part-target_ref'><code>rpptx::Part$target_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XmlPart" data-id="drop_rel"><a href='../../rpptx/html/XmlPart.html#method-XmlPart-drop_rel'><code>rpptx::XmlPart$drop_rel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XmlPart" data-id="initialize"><a href='../../rpptx/html/XmlPart.html#method-XmlPart-initialize'><code>rpptx::XmlPart$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartPart-load"></a>}}
\if{latex}{\out{\hypertarget{method-ChartPart-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartPart$load(partname, content_type, package, blob)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartPart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChartPart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartPart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
