% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-chart.R
\name{ChartPart_new}
\alias{ChartPart_new}
\title{Create a new ChartPart containing \code{chart_type} data from \code{chart_data}}
\usage{
ChartPart_new(chart_type, chart_data, package)
}
\arguments{
\item{chart_type}{An integer from XL_CHART_TYPE.}

\item{chart_data}{A CategoryChartData, XyChartData, or BubbleChartData.}

\item{package}{The OPC package.}
}
\value{
A ChartPart.
}
\description{
Create a new ChartPart containing \code{chart_type} data from \code{chart_data}
}
\keyword{internal}
