% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{ChartTitle}
\alias{ChartTitle}
\title{Chart title domain object}
\description{
Chart title domain object

Chart title domain object
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChartTitle-new}{\code{ChartTitle$new()}}
\item \href{#method-ChartTitle-clone}{\code{ChartTitle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartTitle-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChartTitle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartTitle$new(title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartTitle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChartTitle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartTitle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
