% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-chart.R
\name{ChartWorkbook}
\alias{ChartWorkbook}
\title{Manages the embedded Excel workbook for a chart}
\description{
Manages the embedded Excel workbook for a chart

Manages the embedded Excel workbook for a chart
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChartWorkbook-new}{\code{ChartWorkbook$new()}}
\item \href{#method-ChartWorkbook-update_from_xlsx_blob}{\code{ChartWorkbook$update_from_xlsx_blob()}}
\item \href{#method-ChartWorkbook-clone}{\code{ChartWorkbook$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartWorkbook-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChartWorkbook-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartWorkbook$new(chart_space_elm, chart_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartWorkbook-update_from_xlsx_blob"></a>}}
\if{latex}{\out{\hypertarget{method-ChartWorkbook-update_from_xlsx_blob}{}}}
\subsection{Method \code{update_from_xlsx_blob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartWorkbook$update_from_xlsx_blob(xlsx_blob)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChartWorkbook-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChartWorkbook-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartWorkbook$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
