% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-ns.R
\name{clark_to_nsptag}
\alias{clark_to_nsptag}
\title{Convert a Clark-notation tag to a namespace-prefixed tag}
\usage{
clark_to_nsptag(clark_name)
}
\arguments{
\item{clark_name}{A Clark-notation tag, e.g.
\code{"{http://schemas.openxmlformats.org/presentationml/2006/main}cSld"}.}
}
\value{
A namespace-prefixed tag string, e.g. \code{"p:cSld"}.
}
\description{
Convert a Clark-notation tag to a namespace-prefixed tag
}
\keyword{internal}
