% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-color.R
\name{ColorFormat}
\alias{ColorFormat}
\title{Color format accessor}
\description{
Wraps the "color choice parent" element (e.g. \verb{<a:solidFill>}, \verb{<a:fgClr>},
\verb{<a:bgClr>}, or a gradient stop \verb{<a:gs>}) and exposes color properties.
}
\details{
Access via \code{FillFormat$fore_color}, \code{FillFormat$back_color}, or
\code{LineFormat$color}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ColorFormat-new}{\code{ColorFormat$new()}}
\item \href{#method-ColorFormat-clone}{\code{ColorFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-ColorFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorFormat$new(xPr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xPr}}{A ColorChoiceParent element (e.g. CT_SolidColorFillProperties).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ColorFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
