% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{ContentTypeMap}
\alias{ContentTypeMap}
\title{Resolves partname to content type from [Content_Types].xml}
\description{
Resolves partname to content type from [Content_Types].xml

Resolves partname to content type from [Content_Types].xml
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContentTypeMap-new}{\code{ContentTypeMap$new()}}
\item \href{#method-ContentTypeMap-get}{\code{ContentTypeMap$get()}}
\item \href{#method-ContentTypeMap-clone}{\code{ContentTypeMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContentTypeMap-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContentTypeMap-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentTypeMap$new(overrides, defaults)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContentTypeMap-get"></a>}}
\if{latex}{\out{\hypertarget{method-ContentTypeMap-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentTypeMap$get(partname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContentTypeMap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContentTypeMap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentTypeMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
