% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-coreprops.R
\name{CorePropertiesPart_default}
\alias{CorePropertiesPart_default}
\title{Create a default CorePropertiesPart}
\usage{
CorePropertiesPart_default(package)
}
\description{
Create a default CorePropertiesPart
}
\keyword{internal}
