% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-constants.R
\docType{data}
\name{CT}
\alias{CT}
\title{Content type URIs (like MIME-types) that specify a part's format.}
\format{
An object of class \code{list} of length 104.
}
\usage{
CT
}
\description{
Content type URIs (like MIME-types) that specify a part's format.
}
\keyword{internal}
