% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_Connector_new_cxnSp}
\alias{CT_Connector_new_cxnSp}
\title{Create a new <p:cxnSp> connector element}
\usage{
CT_Connector_new_cxnSp(
  shape_id,
  name,
  prst,
  x,
  y,
  cx,
  cy,
  flipH = FALSE,
  flipV = FALSE
)
}
\description{
Create a new <p:cxnSp> connector element
}
\keyword{internal}
