% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-coreprops.R
\name{CT_CoreProperties}
\alias{CT_CoreProperties}
\title{CT_CoreProperties XML element}
\description{
CT_CoreProperties XML element

CT_CoreProperties XML element
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{CT_CoreProperties}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_CoreProperties-author_text}{\code{CT_CoreProperties$author_text()}}
\item \href{#method-CT_CoreProperties-category_text}{\code{CT_CoreProperties$category_text()}}
\item \href{#method-CT_CoreProperties-comments_text}{\code{CT_CoreProperties$comments_text()}}
\item \href{#method-CT_CoreProperties-contentStatus_text}{\code{CT_CoreProperties$contentStatus_text()}}
\item \href{#method-CT_CoreProperties-identifier_text}{\code{CT_CoreProperties$identifier_text()}}
\item \href{#method-CT_CoreProperties-keywords_text}{\code{CT_CoreProperties$keywords_text()}}
\item \href{#method-CT_CoreProperties-language_text}{\code{CT_CoreProperties$language_text()}}
\item \href{#method-CT_CoreProperties-lastModifiedBy_text}{\code{CT_CoreProperties$lastModifiedBy_text()}}
\item \href{#method-CT_CoreProperties-subject_text}{\code{CT_CoreProperties$subject_text()}}
\item \href{#method-CT_CoreProperties-title_text}{\code{CT_CoreProperties$title_text()}}
\item \href{#method-CT_CoreProperties-version_text}{\code{CT_CoreProperties$version_text()}}
\item \href{#method-CT_CoreProperties-created_datetime}{\code{CT_CoreProperties$created_datetime()}}
\item \href{#method-CT_CoreProperties-lastPrinted_datetime}{\code{CT_CoreProperties$lastPrinted_datetime()}}
\item \href{#method-CT_CoreProperties-modified_datetime}{\code{CT_CoreProperties$modified_datetime()}}
\item \href{#method-CT_CoreProperties-revision_number}{\code{CT_CoreProperties$revision_number()}}
\item \href{#method-CT_CoreProperties-clone}{\code{CT_CoreProperties$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-author_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-author_text}{}}}
\subsection{Method \code{author_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$author_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-category_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-category_text}{}}}
\subsection{Method \code{category_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$category_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-comments_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-comments_text}{}}}
\subsection{Method \code{comments_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$comments_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-contentStatus_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-contentStatus_text}{}}}
\subsection{Method \code{contentStatus_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$contentStatus_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-identifier_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-identifier_text}{}}}
\subsection{Method \code{identifier_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$identifier_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-keywords_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-keywords_text}{}}}
\subsection{Method \code{keywords_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$keywords_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-language_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-language_text}{}}}
\subsection{Method \code{language_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$language_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-lastModifiedBy_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-lastModifiedBy_text}{}}}
\subsection{Method \code{lastModifiedBy_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$lastModifiedBy_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-subject_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-subject_text}{}}}
\subsection{Method \code{subject_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$subject_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-title_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-title_text}{}}}
\subsection{Method \code{title_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$title_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-version_text"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-version_text}{}}}
\subsection{Method \code{version_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$version_text(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-created_datetime"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-created_datetime}{}}}
\subsection{Method \code{created_datetime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$created_datetime(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-lastPrinted_datetime"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-lastPrinted_datetime}{}}}
\subsection{Method \code{lastPrinted_datetime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$lastPrinted_datetime(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-modified_datetime"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-modified_datetime}{}}}
\subsection{Method \code{modified_datetime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$modified_datetime(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-revision_number"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-revision_number}{}}}
\subsection{Method \code{revision_number()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$revision_number(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_CoreProperties-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_CoreProperties-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_CoreProperties$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
