% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{CT_DateAx}
\alias{CT_DateAx}
\title{Date axis XML element}
\description{
Date axis XML element

Date axis XML element
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{\link[rpptx:BaseAxisElement]{rpptx::BaseAxisElement}} -> \code{CT_DateAx}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_DateAx-clone}{\code{CT_DateAx$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".add_majorTickMark"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.add_majorTickMark'><code>rpptx::BaseAxisElement$.add_majorTickMark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".add_minorTickMark"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.add_minorTickMark'><code>rpptx::BaseAxisElement$.add_minorTickMark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".remove_majorGridlines"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.remove_majorGridlines'><code>rpptx::BaseAxisElement$.remove_majorGridlines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".remove_majorTickMark"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.remove_majorTickMark'><code>rpptx::BaseAxisElement$.remove_majorTickMark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".remove_minorGridlines"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.remove_minorGridlines'><code>rpptx::BaseAxisElement$.remove_minorGridlines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".remove_minorTickMark"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.remove_minorTickMark'><code>rpptx::BaseAxisElement$.remove_minorTickMark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id=".remove_title"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-.remove_title'><code>rpptx::BaseAxisElement$.remove_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_delete_"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_delete_'><code>rpptx::BaseAxisElement$get_or_add_delete_()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_majorGridlines"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_majorGridlines'><code>rpptx::BaseAxisElement$get_or_add_majorGridlines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_minorGridlines"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_minorGridlines'><code>rpptx::BaseAxisElement$get_or_add_minorGridlines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_numFmt"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_numFmt'><code>rpptx::BaseAxisElement$get_or_add_numFmt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_scaling"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_scaling'><code>rpptx::BaseAxisElement$get_or_add_scaling()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_spPr"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_spPr'><code>rpptx::BaseAxisElement$get_or_add_spPr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_tickLblPos"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_tickLblPos'><code>rpptx::BaseAxisElement$get_or_add_tickLblPos()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_title"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_title'><code>rpptx::BaseAxisElement$get_or_add_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxisElement" data-id="get_or_add_txPr"><a href='../../rpptx/html/BaseAxisElement.html#method-BaseAxisElement-get_or_add_txPr'><code>rpptx::BaseAxisElement$get_or_add_txPr()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DateAx-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DateAx-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DateAx$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
