% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{CT_DLbl}
\alias{CT_DLbl}
\title{Individual data label element}
\description{
Individual data label element

Individual data label element
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{CT_DLbl}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_DLbl-get_or_add_tx}{\code{CT_DLbl$get_or_add_tx()}}
\item \href{#method-CT_DLbl-get_or_add_txPr}{\code{CT_DLbl$get_or_add_txPr()}}
\item \href{#method-CT_DLbl-get_or_add_dLblPos}{\code{CT_DLbl$get_or_add_dLblPos()}}
\item \href{#method-CT_DLbl-get_or_add_rich}{\code{CT_DLbl$get_or_add_rich()}}
\item \href{#method-CT_DLbl-get_or_add_tx_rich}{\code{CT_DLbl$get_or_add_tx_rich()}}
\item \href{#method-CT_DLbl-remove_tx_rich}{\code{CT_DLbl$remove_tx_rich()}}
\item \href{#method-CT_DLbl-.remove_spPr}{\code{CT_DLbl$.remove_spPr()}}
\item \href{#method-CT_DLbl-.remove_txPr}{\code{CT_DLbl$.remove_txPr()}}
\item \href{#method-CT_DLbl-.remove_dLblPos}{\code{CT_DLbl$.remove_dLblPos()}}
\item \href{#method-CT_DLbl-clone}{\code{CT_DLbl$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-get_or_add_tx"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-get_or_add_tx}{}}}
\subsection{Method \code{get_or_add_tx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$get_or_add_tx()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-get_or_add_txPr"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-get_or_add_txPr}{}}}
\subsection{Method \code{get_or_add_txPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$get_or_add_txPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-get_or_add_dLblPos"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-get_or_add_dLblPos}{}}}
\subsection{Method \code{get_or_add_dLblPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$get_or_add_dLblPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-get_or_add_rich"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-get_or_add_rich}{}}}
\subsection{Method \code{get_or_add_rich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$get_or_add_rich()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-get_or_add_tx_rich"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-get_or_add_tx_rich}{}}}
\subsection{Method \code{get_or_add_tx_rich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$get_or_add_tx_rich()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-remove_tx_rich"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-remove_tx_rich}{}}}
\subsection{Method \code{remove_tx_rich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$remove_tx_rich()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-.remove_spPr"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-.remove_spPr}{}}}
\subsection{Method \code{.remove_spPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$.remove_spPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-.remove_txPr"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-.remove_txPr}{}}}
\subsection{Method \code{.remove_txPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$.remove_txPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-.remove_dLblPos"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-.remove_dLblPos}{}}}
\subsection{Method \code{.remove_dLblPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$.remove_dLblPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
