% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{CT_DLbls}
\alias{CT_DLbls}
\title{Data labels container element}
\description{
Data labels container element

Data labels container element
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{CT_DLbls}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_DLbls-get_or_add_txPr}{\code{CT_DLbls$get_or_add_txPr()}}
\item \href{#method-CT_DLbls-get_or_add_numFmt}{\code{CT_DLbls$get_or_add_numFmt()}}
\item \href{#method-CT_DLbls-get_or_add_dLblPos}{\code{CT_DLbls$get_or_add_dLblPos()}}
\item \href{#method-CT_DLbls-.remove_dLblPos}{\code{CT_DLbls$.remove_dLblPos()}}
\item \href{#method-CT_DLbls-get_or_add_showCatName}{\code{CT_DLbls$get_or_add_showCatName()}}
\item \href{#method-CT_DLbls-get_or_add_showLegendKey}{\code{CT_DLbls$get_or_add_showLegendKey()}}
\item \href{#method-CT_DLbls-get_or_add_showPercent}{\code{CT_DLbls$get_or_add_showPercent()}}
\item \href{#method-CT_DLbls-get_or_add_showSerName}{\code{CT_DLbls$get_or_add_showSerName()}}
\item \href{#method-CT_DLbls-get_or_add_showVal}{\code{CT_DLbls$get_or_add_showVal()}}
\item \href{#method-CT_DLbls-get_dLbl_for_point}{\code{CT_DLbls$get_dLbl_for_point()}}
\item \href{#method-CT_DLbls-get_or_add_dLbl_for_point}{\code{CT_DLbls$get_or_add_dLbl_for_point()}}
\item \href{#method-CT_DLbls-.insert_dLbl_in_sequence}{\code{CT_DLbls$.insert_dLbl_in_sequence()}}
\item \href{#method-CT_DLbls-clone}{\code{CT_DLbls$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_txPr"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_txPr}{}}}
\subsection{Method \code{get_or_add_txPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_txPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_numFmt"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_numFmt}{}}}
\subsection{Method \code{get_or_add_numFmt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_numFmt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_dLblPos"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_dLblPos}{}}}
\subsection{Method \code{get_or_add_dLblPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_dLblPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-.remove_dLblPos"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-.remove_dLblPos}{}}}
\subsection{Method \code{.remove_dLblPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$.remove_dLblPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_showCatName"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_showCatName}{}}}
\subsection{Method \code{get_or_add_showCatName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_showCatName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_showLegendKey"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_showLegendKey}{}}}
\subsection{Method \code{get_or_add_showLegendKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_showLegendKey()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_showPercent"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_showPercent}{}}}
\subsection{Method \code{get_or_add_showPercent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_showPercent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_showSerName"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_showSerName}{}}}
\subsection{Method \code{get_or_add_showSerName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_showSerName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_showVal"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_showVal}{}}}
\subsection{Method \code{get_or_add_showVal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_showVal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_dLbl_for_point"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_dLbl_for_point}{}}}
\subsection{Method \code{get_dLbl_for_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_dLbl_for_point(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-get_or_add_dLbl_for_point"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-get_or_add_dLbl_for_point}{}}}
\subsection{Method \code{get_or_add_dLbl_for_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$get_or_add_dLbl_for_point(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-.insert_dLbl_in_sequence"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-.insert_dLbl_in_sequence}{}}}
\subsection{Method \code{.insert_dLbl_in_sequence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$.insert_dLbl_in_sequence(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_DLbls-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_DLbls-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_DLbls$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
