% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_GraphicalObjectFrame_new_chart_graphicFrame}
\alias{CT_GraphicalObjectFrame_new_chart_graphicFrame}
\title{Create a new <p:graphicFrame> referencing a chart part via rId}
\usage{
CT_GraphicalObjectFrame_new_chart_graphicFrame(id, name, rId, x, y, cx, cy)
}
\description{
Create a new <p:graphicFrame> referencing a chart part via rId
}
\keyword{internal}
