% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-table.R
\name{CT_GraphicalObjectFrame_new_table_graphicFrame}
\alias{CT_GraphicalObjectFrame_new_table_graphicFrame}
\title{Create a new <p:graphicFrame> wrapping a table}
\usage{
CT_GraphicalObjectFrame_new_table_graphicFrame(
  id,
  name,
  rows,
  cols,
  x,
  y,
  cx,
  cy
)
}
\description{
Create a new <p:graphicFrame> wrapping a table
}
\keyword{internal}
