% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_GroupShape_new_grpSp}
\alias{CT_GroupShape_new_grpSp}
\title{Create a new empty <p:grpSp> with bounding-box transform}
\usage{
CT_GroupShape_new_grpSp(id, name, x, y, cx, cy)
}
\description{
Create a new empty <p:grpSp> with bounding-box transform
}
\keyword{internal}
