% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_Picture_new_pic}
\alias{CT_Picture_new_pic}
\title{Create a new <p:pic> picture element}
\usage{
CT_Picture_new_pic(shape_id, name, desc, rId, x, y, cx, cy)
}
\description{
Create a new <p:pic> picture element
}
\keyword{internal}
