% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{CT_Scaling}
\alias{CT_Scaling}
\title{Axis scaling element}
\description{
Axis scaling element

Axis scaling element
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{CT_Scaling}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_Scaling-.remove_max}{\code{CT_Scaling$.remove_max()}}
\item \href{#method-CT_Scaling-.add_max}{\code{CT_Scaling$.add_max()}}
\item \href{#method-CT_Scaling-.remove_min}{\code{CT_Scaling$.remove_min()}}
\item \href{#method-CT_Scaling-.add_min}{\code{CT_Scaling$.add_min()}}
\item \href{#method-CT_Scaling-get_or_add_orientation}{\code{CT_Scaling$get_or_add_orientation()}}
\item \href{#method-CT_Scaling-.remove_orientation}{\code{CT_Scaling$.remove_orientation()}}
\item \href{#method-CT_Scaling-clone}{\code{CT_Scaling$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-.remove_max"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-.remove_max}{}}}
\subsection{Method \code{.remove_max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$.remove_max()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-.add_max"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-.add_max}{}}}
\subsection{Method \code{.add_max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$.add_max(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-.remove_min"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-.remove_min}{}}}
\subsection{Method \code{.remove_min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$.remove_min()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-.add_min"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-.add_min}{}}}
\subsection{Method \code{.add_min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$.add_min(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-get_or_add_orientation"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-get_or_add_orientation}{}}}
\subsection{Method \code{get_or_add_orientation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$get_or_add_orientation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-.remove_orientation"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-.remove_orientation}{}}}
\subsection{Method \code{.remove_orientation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$.remove_orientation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_Scaling-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_Scaling-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_Scaling$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
