% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{CT_SeriesComposite}
\alias{CT_SeriesComposite}
\title{Chart series element (composite; all series use the same c:ser tag)}
\description{
Chart series element (composite; all series use the same c:ser tag)

Chart series element (composite; all series use the same c:ser tag)
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{CT_SeriesComposite}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_SeriesComposite-get_or_add_dLbls}{\code{CT_SeriesComposite$get_or_add_dLbls()}}
\item \href{#method-CT_SeriesComposite-get_or_add_marker}{\code{CT_SeriesComposite$get_or_add_marker()}}
\item \href{#method-CT_SeriesComposite-get_or_add_invertIfNegative}{\code{CT_SeriesComposite$get_or_add_invertIfNegative()}}
\item \href{#method-CT_SeriesComposite-get_or_add_smooth}{\code{CT_SeriesComposite$get_or_add_smooth()}}
\item \href{#method-CT_SeriesComposite-get_or_add_spPr}{\code{CT_SeriesComposite$get_or_add_spPr()}}
\item \href{#method-CT_SeriesComposite-get_dLbl}{\code{CT_SeriesComposite$get_dLbl()}}
\item \href{#method-CT_SeriesComposite-get_or_add_dLbl}{\code{CT_SeriesComposite$get_or_add_dLbl()}}
\item \href{#method-CT_SeriesComposite-get_or_add_dPt_for_point}{\code{CT_SeriesComposite$get_or_add_dPt_for_point()}}
\item \href{#method-CT_SeriesComposite-clone}{\code{CT_SeriesComposite$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_dLbls"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_dLbls}{}}}
\subsection{Method \code{get_or_add_dLbls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_dLbls()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_marker"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_marker}{}}}
\subsection{Method \code{get_or_add_marker()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_marker()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_invertIfNegative"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_invertIfNegative}{}}}
\subsection{Method \code{get_or_add_invertIfNegative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_invertIfNegative()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_smooth"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_smooth}{}}}
\subsection{Method \code{get_or_add_smooth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_smooth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_spPr"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_spPr}{}}}
\subsection{Method \code{get_or_add_spPr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_spPr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_dLbl"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_dLbl}{}}}
\subsection{Method \code{get_dLbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_dLbl(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_dLbl"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_dLbl}{}}}
\subsection{Method \code{get_or_add_dLbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_dLbl(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-get_or_add_dPt_for_point"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-get_or_add_dPt_for_point}{}}}
\subsection{Method \code{get_or_add_dPt_for_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$get_or_add_dPt_for_point(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_SeriesComposite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_SeriesComposite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_SeriesComposite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
