% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_Shape_new_autoshape_sp}
\alias{CT_Shape_new_autoshape_sp}
\title{Create a new autoshape <p:sp> element}
\usage{
CT_Shape_new_autoshape_sp(id, name, prst, x, y, cx, cy)
}
\description{
Create a new autoshape <p:sp> element
}
\keyword{internal}
