% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_Shape_new_freeform_sp}
\alias{CT_Shape_new_freeform_sp}
\title{Create a new freeform <p:sp> element with custom geometry}
\usage{
CT_Shape_new_freeform_sp(id, name, x, y, cx, cy)
}
\description{
Create a new freeform <p:sp> element with custom geometry
}
\keyword{internal}
