% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_Shape_new_placeholder_sp}
\alias{CT_Shape_new_placeholder_sp}
\title{Create a new placeholder <p:sp> element}
\usage{
CT_Shape_new_placeholder_sp(id, name, ph_type, orient, sz, idx)
}
\description{
Create a new placeholder <p:sp> element
}
\keyword{internal}
