% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_Shape_new_textbox_sp}
\alias{CT_Shape_new_textbox_sp}
\title{Create a new textbox <p:sp> element}
\usage{
CT_Shape_new_textbox_sp(id, name, x, y, cx, cy)
}
\description{
Create a new textbox <p:sp> element
}
\keyword{internal}
