% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-chart.R
\name{CT_xChart}
\alias{CT_xChart}
\title{Generic xChart element (barChart, lineChart, etc.)}
\description{
Generic xChart element (barChart, lineChart, etc.)

Generic xChart element (barChart, lineChart, etc.)
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{CT_xChart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_xChart-iter_sers}{\code{CT_xChart$iter_sers()}}
\item \href{#method-CT_xChart-.remove_dLbls}{\code{CT_xChart$.remove_dLbls()}}
\item \href{#method-CT_xChart-add_dLbls_default}{\code{CT_xChart$add_dLbls_default()}}
\item \href{#method-CT_xChart-get_or_add_varyColors}{\code{CT_xChart$get_or_add_varyColors()}}
\item \href{#method-CT_xChart-get_or_add_gapWidth}{\code{CT_xChart$get_or_add_gapWidth()}}
\item \href{#method-CT_xChart-get_or_add_overlap}{\code{CT_xChart$get_or_add_overlap()}}
\item \href{#method-CT_xChart-.remove_overlap}{\code{CT_xChart$.remove_overlap()}}
\item \href{#method-CT_xChart-clone}{\code{CT_xChart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-iter_sers"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-iter_sers}{}}}
\subsection{Method \code{iter_sers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$iter_sers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-.remove_dLbls"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-.remove_dLbls}{}}}
\subsection{Method \code{.remove_dLbls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$.remove_dLbls()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-add_dLbls_default"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-add_dLbls_default}{}}}
\subsection{Method \code{add_dLbls_default()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$add_dLbls_default()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-get_or_add_varyColors"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-get_or_add_varyColors}{}}}
\subsection{Method \code{get_or_add_varyColors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$get_or_add_varyColors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-get_or_add_gapWidth"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-get_or_add_gapWidth}{}}}
\subsection{Method \code{get_or_add_gapWidth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$get_or_add_gapWidth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-get_or_add_overlap"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-get_or_add_overlap}{}}}
\subsection{Method \code{get_or_add_overlap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$get_or_add_overlap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-.remove_overlap"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-.remove_overlap}{}}}
\subsection{Method \code{.remove_overlap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$.remove_overlap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_xChart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_xChart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_xChart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
