% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{DataLabel}
\alias{DataLabel}
\title{Individual data point label object}
\description{
Individual data point label object

Individual data point label object
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataLabel-new}{\code{DataLabel$new()}}
\item \href{#method-DataLabel-.get_or_add_dLbl}{\code{DataLabel$.get_or_add_dLbl()}}
\item \href{#method-DataLabel-.get_or_add_rich}{\code{DataLabel$.get_or_add_rich()}}
\item \href{#method-DataLabel-clone}{\code{DataLabel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataLabel-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataLabel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataLabel$new(ser, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataLabel-.get_or_add_dLbl"></a>}}
\if{latex}{\out{\hypertarget{method-DataLabel-.get_or_add_dLbl}{}}}
\subsection{Method \code{.get_or_add_dLbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataLabel$.get_or_add_dLbl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataLabel-.get_or_add_rich"></a>}}
\if{latex}{\out{\hypertarget{method-DataLabel-.get_or_add_rich}{}}}
\subsection{Method \code{.get_or_add_rich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataLabel$.get_or_add_rich()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataLabel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataLabel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataLabel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
