% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{DataLabels}
\alias{DataLabels}
\title{Data labels collection object}
\description{
Controls display properties of all data labels in a plot or series.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataLabels-new}{\code{DataLabels$new()}}
\item \href{#method-DataLabels-clone}{\code{DataLabels$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataLabels-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataLabels-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataLabels$new(dLbls)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataLabels-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataLabels-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataLabels$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
