% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-serialized.R
\name{.content_types_xml_for}
\alias{.content_types_xml_for}
\title{Generate [Content_Types].xml bytes for a set of parts}
\usage{
.content_types_xml_for(parts)
}
\arguments{
\item{parts}{A list of Part objects.}
}
\value{
Raw bytes of the content types XML.
}
\description{
Generate [Content_Types].xml bytes for a set of parts
}
\keyword{internal}
