% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-dml-fill.R
\docType{data}
\name{.fill_choice_tags}
\alias{.fill_choice_tags}
\title{Tags for fill choice elements within spPr/ln (Clark notation)}
\format{
An object of class \code{character} of length 6.
}
\usage{
.fill_choice_tags
}
\description{
Tags for fill choice elements within spPr/ln (Clark notation)
}
\keyword{internal}
