% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-serialized.R
\name{.read_zip_blobs}
\alias{.read_zip_blobs}
\title{Read all blobs from a ZIP file}
\usage{
.read_zip_blobs(pkg_file)
}
\arguments{
\item{pkg_file}{Path to a .pptx file or a raw connection.}
}
\value{
Named list mapping PackURI strings to raw vectors.
}
\description{
Read all blobs from a ZIP file
}
\keyword{internal}
