% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-serialized.R
\name{.write_zip_package}
\alias{.write_zip_package}
\title{Write items to a ZIP file}
\usage{
.write_zip_package(pkg_file, items)
}
\arguments{
\item{pkg_file}{Path to output file.}

\item{items}{A list of lists with \code{membername} and \code{blob} elements.}
}
\description{
Write items to a ZIP file
}
\keyword{internal}
