% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{ElementProxy}
\alias{ElementProxy}
\title{Base proxy wrapping an XML element}
\description{
Base proxy wrapping an XML element

Base proxy wrapping an XML element
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementProxy-new}{\code{ElementProxy$new()}}
\item \href{#method-ElementProxy-clone}{\code{ElementProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProxy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProxy$new(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
