% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-chart.R
\name{EmbeddedXlsxPart_new}
\alias{EmbeddedXlsxPart_new}
\title{Create a new EmbeddedXlsxPart containing \code{blob}}
\usage{
EmbeddedXlsxPart_new(blob, package)
}
\arguments{
\item{blob}{Raw bytes (the xlsx file).}

\item{package}{The OPC package.}
}
\value{
An EmbeddedXlsxPart.
}
\description{
Create a new EmbeddedXlsxPart containing \code{blob}
}
\keyword{internal}
