% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-fill.R
\name{FillFormat}
\alias{FillFormat}
\title{Fill format accessor}
\description{
Wraps the shape-properties element (\verb{<p:spPr>}) and exposes fill properties.
Access via \code{shape$fill}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FillFormat-new}{\code{FillFormat$new()}}
\item \href{#method-FillFormat-solid}{\code{FillFormat$solid()}}
\item \href{#method-FillFormat-background}{\code{FillFormat$background()}}
\item \href{#method-FillFormat-gradient}{\code{FillFormat$gradient()}}
\item \href{#method-FillFormat-patterned}{\code{FillFormat$patterned()}}
\item \href{#method-FillFormat-clone}{\code{FillFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FillFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-FillFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FillFormat$new(spPr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spPr}}{A CT_ShapeProperties or similar element that contains fill.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FillFormat-solid"></a>}}
\if{latex}{\out{\hypertarget{method-FillFormat-solid}{}}}
\subsection{Method \code{solid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FillFormat$solid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FillFormat-background"></a>}}
\if{latex}{\out{\hypertarget{method-FillFormat-background}{}}}
\subsection{Method \code{background()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FillFormat$background()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FillFormat-gradient"></a>}}
\if{latex}{\out{\hypertarget{method-FillFormat-gradient}{}}}
\subsection{Method \code{gradient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FillFormat$gradient()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FillFormat-patterned"></a>}}
\if{latex}{\out{\hypertarget{method-FillFormat-patterned}{}}}
\subsection{Method \code{patterned()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FillFormat$patterned()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FillFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FillFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FillFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
