% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-text.R
\name{Font}
\alias{Font}
\title{Character properties for a run, paragraph-default, or end-of-paragraph.}
\description{
Wraps an \code{a:rPr}, \code{a:defRPr}, or \code{a:endParaRPr} element. Provides access
to font name, size, bold, italic, and underline. All properties are R/W;
assigning NULL removes the override and inherits from the style hierarchy.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Font-new}{\code{Font$new()}}
\item \href{#method-Font-element}{\code{Font$element()}}
\item \href{#method-Font-clone}{\code{Font$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Font-new"></a>}}
\if{latex}{\out{\hypertarget{method-Font-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$new(rPr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Font-element"></a>}}
\if{latex}{\out{\hypertarget{method-Font-element}{}}}
\subsection{Method \code{element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Font-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Font-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
