% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_functions.R
\name{get_slide_id}
\alias{get_slide_id}
\title{Get the unique slide ID for a slide using the slide number}
\usage{
get_slide_id(pres, slide_num)
}
\arguments{
\item{pres}{Python pptx Presentation object.}

\item{slide_num}{(int) The slide number of the target slide.}
}
\value{

}
\description{
Get the unique slide ID for a slide using the slide number
}
