% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-fill.R
\name{GradientStop}
\alias{GradientStop}
\title{Gradient stop proxy}
\description{
Gradient stop proxy

Gradient stop proxy
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GradientStop-new}{\code{GradientStop$new()}}
\item \href{#method-GradientStop-clone}{\code{GradientStop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradientStop-new"></a>}}
\if{latex}{\out{\hypertarget{method-GradientStop-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientStop$new(gs_elm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradientStop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GradientStop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientStop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
