% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-fill.R
\name{GradientStops}
\alias{GradientStops}
\title{Gradient stops collection}
\description{
Gradient stops collection

Gradient stops collection
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GradientStops-new}{\code{GradientStops$new()}}
\item \href{#method-GradientStops-get}{\code{GradientStops$get()}}
\item \href{#method-GradientStops-to_list}{\code{GradientStops$to_list()}}
\item \href{#method-GradientStops-clone}{\code{GradientStops$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradientStops-new"></a>}}
\if{latex}{\out{\hypertarget{method-GradientStops-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientStops$new(gsLst_elm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradientStops-get"></a>}}
\if{latex}{\out{\hypertarget{method-GradientStops-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientStops$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradientStops-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GradientStops-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientStops$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradientStops-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GradientStops-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientStops$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
