% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{GroupShapes}
\alias{GroupShapes}
\title{Shape collection for a group shape}
\description{
Shape collection for a group shape

Shape collection for a group shape
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{GroupShapes}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupShapes-new}{\code{GroupShapes$new()}}
\item \href{#method-GroupShapes-get}{\code{GroupShapes$get()}}
\item \href{#method-GroupShapes-to_list}{\code{GroupShapes$to_list()}}
\item \href{#method-GroupShapes-clone}{\code{GroupShapes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupShapes-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupShapes-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupShapes$new(grpSp, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupShapes-get"></a>}}
\if{latex}{\out{\hypertarget{method-GroupShapes-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupShapes$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupShapes-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GroupShapes-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupShapes$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupShapes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupShapes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupShapes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
