% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-text.R
\name{Hyperlink}
\alias{Hyperlink}
\title{Hyperlink on a text run}
\description{
Provides read/write access to the URL of a hyperlink on an \code{a:r} element.
Access via \code{run$hyperlink}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Hyperlink-new}{\code{Hyperlink$new()}}
\item \href{#method-Hyperlink-clone}{\code{Hyperlink$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hyperlink-new"></a>}}
\if{latex}{\out{\hypertarget{method-Hyperlink-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hyperlink$new(rPr, slide_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hyperlink-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Hyperlink-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hyperlink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
