% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-image.R
\name{Image}
\alias{Image}
\title{Image value object}
\description{
Immutable wrapper around raw image bytes that exposes ext, content_type,
sha1, size (pixels) and dpi.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Image-new}{\code{Image$new()}}
\item \href{#method-Image-sha1}{\code{Image$sha1()}}
\item \href{#method-Image-clone}{\code{Image$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Image-new"></a>}}
\if{latex}{\out{\hypertarget{method-Image-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Image$new(blob, filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Image-sha1"></a>}}
\if{latex}{\out{\hypertarget{method-Image-sha1}{}}}
\subsection{Method \code{sha1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Image$sha1()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Image-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Image-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Image$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
