% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-image.R
\name{ImagePart}
\alias{ImagePart}
\title{Image OPC part}
\description{
Stores the raw bytes of an image and provides dimension scaling.
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:Part]{rpptx::Part}} -> \code{ImagePart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImagePart-new}{\code{ImagePart$new()}}
\item \href{#method-ImagePart-scale}{\code{ImagePart$scale()}}
\item \href{#method-ImagePart-clone}{\code{ImagePart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="drop_rel"><a href='../../rpptx/html/Part.html#method-Part-drop_rel'><code>rpptx::Part$drop_rel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="has_rels"><a href='../../rpptx/html/Part.html#method-Part-has_rels'><code>rpptx::Part$has_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="load_rels_from_xml"><a href='../../rpptx/html/Part.html#method-Part-load_rels_from_xml'><code>rpptx::Part$load_rels_from_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="part_related_by"><a href='../../rpptx/html/Part.html#method-Part-part_related_by'><code>rpptx::Part$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="relate_to"><a href='../../rpptx/html/Part.html#method-Part-relate_to'><code>rpptx::Part$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="related_part"><a href='../../rpptx/html/Part.html#method-Part-related_part'><code>rpptx::Part$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="rels_xml_bytes"><a href='../../rpptx/html/Part.html#method-Part-rels_xml_bytes'><code>rpptx::Part$rels_xml_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="target_ref"><a href='../../rpptx/html/Part.html#method-Part-target_ref'><code>rpptx::Part$target_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagePart-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImagePart-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagePart$new(partname, content_type, package, blob, filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagePart-scale"></a>}}
\if{latex}{\out{\hypertarget{method-ImagePart-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagePart$scale(cx, cy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagePart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImagePart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagePart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
