% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{invalid_xml_error}
\alias{invalid_xml_error}
\title{Signal an "invalid XML" error}
\usage{
invalid_xml_error(message)
}
\arguments{
\item{message}{Description of the invalid XML.}
}
\description{
Raised when a value in the XML is not valid according to the schema.
}
\keyword{internal}
