% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-spec.R
\name{is_default_content_type}
\alias{is_default_content_type}
\title{Check if a (ext, content_type) pair is a default content type}
\usage{
is_default_content_type(ext, content_type)
}
\arguments{
\item{ext}{File extension (lowercase).}

\item{content_type}{Content type string.}
}
\value{
Logical.
}
\description{
Check if a (ext, content_type) pair is a default content type
}
\keyword{internal}
