% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{LayoutPlaceholder}
\alias{LayoutPlaceholder}
\title{Placeholder shape on a slide layout}
\description{
Inherits dimensions from the corresponding master placeholder when not
explicitly set on the layout element.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:BaseShape]{rpptx::BaseShape}} -> \code{\link[rpptx:Shape]{rpptx::Shape}} -> \code{LayoutPlaceholder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LayoutPlaceholder-clone}{\code{LayoutPlaceholder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseShape" data-id="initialize"><a href='../../rpptx/html/BaseShape.html#method-BaseShape-initialize'><code>rpptx::BaseShape$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayoutPlaceholder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LayoutPlaceholder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayoutPlaceholder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
