% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{Legend}
\alias{Legend}
\title{Chart legend domain object}
\description{
Chart legend domain object

Chart legend domain object
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Legend-new}{\code{Legend$new()}}
\item \href{#method-Legend-clone}{\code{Legend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Legend-new"></a>}}
\if{latex}{\out{\hypertarget{method-Legend-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Legend$new(legend_elm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Legend-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Legend-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Legend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
