% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Length}
\alias{Length}
\alias{Inches}
\alias{Cm}
\alias{Mm}
\alias{Pt}
\alias{Emu}
\alias{Centipoints}
\title{Create a Length value in English Metric Units (EMU)}
\usage{
Length(emu)

Inches(inches)

Cm(cm)

Mm(mm)

Pt(pt)

Emu(emu)

Centipoints(centipoints)
}
\arguments{
\item{emu}{Integer EMU value.}

\item{inches}{Numeric length in inches.}

\item{cm}{Numeric length in centimeters.}

\item{mm}{Numeric length in millimeters.}

\item{pt}{Numeric length in points.}

\item{centipoints}{Integer length in hundredths of a point (1/7200 inch).}
}
\value{
An integer with class \code{"Length"}.
}
\description{
Base constructor for length values. All length values are stored as integer
EMU values. Use convenience constructors \code{\link[=Inches]{Inches()}}, \code{\link[=Cm]{Cm()}}, \code{\link[=Mm]{Mm()}},
\code{\link[=Pt]{Pt()}}, \code{\link[=Emu]{Emu()}}, and \code{\link[=Centipoints]{Centipoints()}} for common units.
}
