% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-line.R
\name{LineFormat}
\alias{LineFormat}
\title{Line format accessor}
\description{
Wraps the shape-properties element (\verb{<p:spPr>}) and exposes border/line
properties. Access via \code{shape$line}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LineFormat-new}{\code{LineFormat$new()}}
\item \href{#method-LineFormat-clone}{\code{LineFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LineFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-LineFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineFormat$new(spPr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spPr}}{A CT_ShapeProperties element.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LineFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LineFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
