% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{MajorGridlines}
\alias{MajorGridlines}
\title{Major gridlines domain object}
\description{
Major gridlines domain object

Major gridlines domain object
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MajorGridlines-new}{\code{MajorGridlines$new()}}
\item \href{#method-MajorGridlines-clone}{\code{MajorGridlines$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MajorGridlines-new"></a>}}
\if{latex}{\out{\hypertarget{method-MajorGridlines-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MajorGridlines$new(xAx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MajorGridlines-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MajorGridlines-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MajorGridlines$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
