% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{Marker}
\alias{Marker}
\title{Data point marker object}
\description{
Controls size and style of markers on line-type chart series.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Marker-new}{\code{Marker$new()}}
\item \href{#method-Marker-clone}{\code{Marker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Marker-new"></a>}}
\if{latex}{\out{\hypertarget{method-Marker-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Marker$new(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Marker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Marker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Marker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
