% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-media.R
\name{MediaPart}
\alias{MediaPart}
\title{Media part (audio/video OPC part)}
\description{
Media part (audio/video OPC part)

Media part (audio/video OPC part)
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:Part]{rpptx::Part}} -> \code{MediaPart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MediaPart-clone}{\code{MediaPart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="drop_rel"><a href='../../rpptx/html/Part.html#method-Part-drop_rel'><code>rpptx::Part$drop_rel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="has_rels"><a href='../../rpptx/html/Part.html#method-Part-has_rels'><code>rpptx::Part$has_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="initialize"><a href='../../rpptx/html/Part.html#method-Part-initialize'><code>rpptx::Part$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="load_rels_from_xml"><a href='../../rpptx/html/Part.html#method-Part-load_rels_from_xml'><code>rpptx::Part$load_rels_from_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="part_related_by"><a href='../../rpptx/html/Part.html#method-Part-part_related_by'><code>rpptx::Part$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="relate_to"><a href='../../rpptx/html/Part.html#method-Part-relate_to'><code>rpptx::Part$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="related_part"><a href='../../rpptx/html/Part.html#method-Part-related_part'><code>rpptx::Part$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="rels_xml_bytes"><a href='../../rpptx/html/Part.html#method-Part-rels_xml_bytes'><code>rpptx::Part$rels_xml_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="target_ref"><a href='../../rpptx/html/Part.html#method-Part-target_ref'><code>rpptx::Part$target_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MediaPart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MediaPart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaPart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
