% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-shapes.R
\docType{data}
\name{MSO}
\alias{MSO}
\title{MSO shape type enum (alias for MSO_SHAPE_TYPE)}
\format{
An object of class \code{list} of length 26.
}
\usage{
MSO
}
\description{
MSO shape type enum (alias for MSO_SHAPE_TYPE)
}
\keyword{internal}
