% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-text.R
\docType{data}
\name{MSO_AUTO_SIZE}
\alias{MSO_AUTO_SIZE}
\title{MSO_AUTO_SIZE — text-box auto-sizing behaviour}
\format{
An object of class \code{list} of length 4.
}
\usage{
MSO_AUTO_SIZE
}
\description{
Integer constants for \verb{a:bodyPr/@autofit} / spAutoFit / normAutofit.
}
\keyword{internal}
