% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-dml.R
\docType{data}
\name{MSO_COLOR_TYPE}
\alias{MSO_COLOR_TYPE}
\title{MSO_COLOR_TYPE — color source type strings}
\format{
An object of class \code{list} of length 6.
}
\usage{
MSO_COLOR_TYPE
}
\description{
MSO_COLOR_TYPE — color source type strings
}
\keyword{internal}
