% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-shapes.R
\docType{data}
\name{MSO_CONNECTOR}
\alias{MSO_CONNECTOR}
\title{MSO connector type enum (alias for MSO_CONNECTOR_TYPE)}
\format{
An object of class \code{list} of length 4.
}
\usage{
MSO_CONNECTOR
}
\description{
MSO connector type enum (alias for MSO_CONNECTOR_TYPE)
}
\keyword{internal}
