% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-dml.R
\docType{data}
\name{MSO_FILL}
\alias{MSO_FILL}
\title{MSO_FILL — fill type strings}
\format{
An object of class \code{list} of length 6.
}
\usage{
MSO_FILL
}
\description{
MSO_FILL — fill type strings
}
\keyword{internal}
