% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-dml.R
\docType{data}
\name{MSO_LINE_DASH_STYLE}
\alias{MSO_LINE_DASH_STYLE}
\title{MSO_LINE_DASH_STYLE — preset dash pattern strings for \verb{<a:prstDash val="..."/>}}
\format{
An object of class \code{list} of length 8.
}
\usage{
MSO_LINE_DASH_STYLE
}
\description{
MSO_LINE_DASH_STYLE — preset dash pattern strings for \verb{<a:prstDash val="..."/>}
}
\keyword{datasets}
