% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-dml.R
\docType{data}
\name{MSO_PATTERN_TYPE}
\alias{MSO_PATTERN_TYPE}
\title{MSO_PATTERN_TYPE — preset pattern fill type strings}
\format{
An object of class \code{list} of length 42.
}
\usage{
MSO_PATTERN_TYPE
}
\description{
MSO_PATTERN_TYPE — preset pattern fill type strings
}
\keyword{internal}
