% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-dml.R
\docType{data}
\name{MSO_THEME_COLOR}
\alias{MSO_THEME_COLOR}
\title{MSO_THEME_COLOR — theme color XML value strings}
\format{
An object of class \code{list} of length 13.
}
\usage{
MSO_THEME_COLOR
}
\description{
String values for the \code{val} attribute on \verb{<a:schemeClr>}. Assign to
\code{color_format$theme_color}.
}
\keyword{datasets}
