% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-text.R
\docType{data}
\name{MSO_VERTICAL_ANCHOR}
\alias{MSO_VERTICAL_ANCHOR}
\title{MSO_VERTICAL_ANCHOR — vertical text anchor}
\format{
An object of class \code{list} of length 4.
}
\usage{
MSO_VERTICAL_ANCHOR
}
\description{
XML string values for \verb{a:bodyPr/@anchor}. Use with \code{text_frame$vertical_anchor}.
}
\keyword{datasets}
