% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-ns.R
\name{namespaces}
\alias{namespaces}
\title{Return a subset of the namespace map}
\usage{
namespaces(...)
}
\arguments{
\item{...}{Namespace prefixes as character strings.}
}
\value{
A named character vector mapping prefixes to URIs.
}
\description{
Return a subset of the namespace map
}
\keyword{internal}
