% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-coreprops.R
\name{new_ct_coreProperties}
\alias{new_ct_coreProperties}
\title{Create a new empty cp:coreProperties element}
\usage{
new_ct_coreProperties()
}
\description{
Create a new empty cp:coreProperties element
}
\keyword{internal}
