% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-oxml.R
\name{new_ct_relationships}
\alias{new_ct_relationships}
\title{Create a new empty \if{html}{\out{<Relationships>}} element}
\usage{
new_ct_relationships()
}
\value{
A CT_Relationships instance.
}
\description{
Create a new empty \if{html}{\out{<Relationships>}} element
}
\keyword{internal}
