% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-oxml.R
\name{new_ct_types}
\alias{new_ct_types}
\title{Create a new empty \if{html}{\out{<Types>}} element}
\usage{
new_ct_types()
}
\value{
A CT_Types instance.
}
\description{
Create a new empty \if{html}{\out{<Types>}} element
}
\keyword{internal}
