% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-slide.R
\name{NotesSlidePart}
\alias{NotesSlidePart}
\title{Notes slide part}
\description{
Notes slide part

Notes slide part
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:Part]{rpptx::Part}} -> \code{\link[rpptx:XmlPart]{rpptx::XmlPart}} -> \code{\link[rpptx:BaseSlidePart]{rpptx::BaseSlidePart}} -> \code{NotesSlidePart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotesSlidePart-clone}{\code{NotesSlidePart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="has_rels"><a href='../../rpptx/html/Part.html#method-Part-has_rels'><code>rpptx::Part$has_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="load_rels_from_xml"><a href='../../rpptx/html/Part.html#method-Part-load_rels_from_xml'><code>rpptx::Part$load_rels_from_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="part_related_by"><a href='../../rpptx/html/Part.html#method-Part-part_related_by'><code>rpptx::Part$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="relate_to"><a href='../../rpptx/html/Part.html#method-Part-relate_to'><code>rpptx::Part$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="related_part"><a href='../../rpptx/html/Part.html#method-Part-related_part'><code>rpptx::Part$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="rels_xml_bytes"><a href='../../rpptx/html/Part.html#method-Part-rels_xml_bytes'><code>rpptx::Part$rels_xml_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="target_ref"><a href='../../rpptx/html/Part.html#method-Part-target_ref'><code>rpptx::Part$target_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XmlPart" data-id="drop_rel"><a href='../../rpptx/html/XmlPart.html#method-XmlPart-drop_rel'><code>rpptx::XmlPart$drop_rel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XmlPart" data-id="initialize"><a href='../../rpptx/html/XmlPart.html#method-XmlPart-initialize'><code>rpptx::XmlPart$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotesSlidePart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotesSlidePart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotesSlidePart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
