% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-constants.R
\docType{data}
\name{NS}
\alias{NS}
\title{Constant values for OPC XML namespaces.}
\format{
An object of class \code{list} of length 5.
}
\usage{
NS
}
\description{
Constant values for OPC XML namespaces.
}
\keyword{internal}
