% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-ns.R
\name{nsdecls}
\alias{nsdecls}
\title{Return namespace declaration strings}
\usage{
nsdecls(...)
}
\arguments{
\item{...}{Namespace prefixes as character strings.}
}
\value{
A single string with xmlns declarations, e.g.
\code{'xmlns:p="http://..."'}.
}
\description{
Return namespace declaration strings
}
\keyword{internal}
