% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-ns.R
\name{nsuri}
\alias{nsuri}
\title{Get the namespace URI for a given prefix}
\usage{
nsuri(nspfx)
}
\arguments{
\item{nspfx}{A namespace prefix string, e.g. \code{"p"}.}
}
\value{
The namespace URI string.
}
\description{
Get the namespace URI for a given prefix
}
\keyword{internal}
