% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-xmlchemy.R
\name{one_and_only_one}
\alias{one_and_only_one}
\title{Define a OneAndOnlyOne child element specification}
\usage{
one_and_only_one(nsptagname)
}
\arguments{
\item{nsptagname}{Namespace-prefixed tag.}
}
\value{
A list describing the child element spec.
}
\description{
Define a OneAndOnlyOne child element specification
}
\keyword{internal}
