% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{OpcPackage}
\alias{OpcPackage}
\title{Open Packaging Convention package}
\description{
A new instance is constructed by calling \code{OpcPackage$open()} with a path
to a package file (.pptx).
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpcPackage-new}{\code{OpcPackage$new()}}
\item \href{#method-OpcPackage-drop_rel}{\code{OpcPackage$drop_rel()}}
\item \href{#method-OpcPackage-iter_parts}{\code{OpcPackage$iter_parts()}}
\item \href{#method-OpcPackage-iter_rels}{\code{OpcPackage$iter_rels()}}
\item \href{#method-OpcPackage-next_partname}{\code{OpcPackage$next_partname()}}
\item \href{#method-OpcPackage-part_related_by}{\code{OpcPackage$part_related_by()}}
\item \href{#method-OpcPackage-relate_to}{\code{OpcPackage$relate_to()}}
\item \href{#method-OpcPackage-related_part}{\code{OpcPackage$related_part()}}
\item \href{#method-OpcPackage-target_ref}{\code{OpcPackage$target_ref()}}
\item \href{#method-OpcPackage-save}{\code{OpcPackage$save()}}
\item \href{#method-OpcPackage-clone}{\code{OpcPackage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$new(pkg_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-drop_rel"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-drop_rel}{}}}
\subsection{Method \code{drop_rel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$drop_rel(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-iter_parts"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-iter_parts}{}}}
\subsection{Method \code{iter_parts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$iter_parts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-iter_rels"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-iter_rels}{}}}
\subsection{Method \code{iter_rels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$iter_rels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-next_partname"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-next_partname}{}}}
\subsection{Method \code{next_partname()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$next_partname(tmpl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-part_related_by"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-part_related_by}{}}}
\subsection{Method \code{part_related_by()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$part_related_by(reltype)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-relate_to"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-relate_to}{}}}
\subsection{Method \code{relate_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$relate_to(target, reltype, is_external = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-related_part"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-related_part}{}}}
\subsection{Method \code{related_part()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$related_part(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-target_ref"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-target_ref}{}}}
\subsection{Method \code{target_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$target_ref(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-save"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$save(pkg_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpcPackage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpcPackage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpcPackage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
