% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{OpcPackage_open}
\alias{OpcPackage_open}
\title{Open an OPC package file}
\usage{
OpcPackage_open(pkg_file)
}
\arguments{
\item{pkg_file}{Path to a .pptx file.}
}
\value{
An OpcPackage instance.
}
\description{
Open an OPC package file
}
\keyword{internal}
