% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-xmlchemy.R
\name{optional_attribute}
\alias{optional_attribute}
\title{Define an OptionalAttribute specification}
\usage{
optional_attribute(attr_name, simple_type, default = NULL)
}
\arguments{
\item{attr_name}{Attribute name (may be namespace-prefixed).}

\item{simple_type}{A list with \code{from_xml} and \code{to_xml} functions.}

\item{default}{Default value when attribute is absent (default NULL).}
}
\value{
A list describing the attribute spec.
}
\description{
Define an OptionalAttribute specification
}
\keyword{internal}
