% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-xmlchemy.R
\name{OxmlElement}
\alias{OxmlElement}
\title{Create a new XML element with the given namespace-prefixed tag}
\usage{
OxmlElement(nsptag_str, nsmap = NULL)
}
\arguments{
\item{nsptag_str}{Namespace-prefixed tag, e.g. \code{"a:tbl"}.}

\item{nsmap}{Optional named character vector of additional namespace mappings.}
}
\value{
A wrapped BaseOxmlElement (or appropriate subclass).
}
\description{
Create a new XML element with the given namespace-prefixed tag
}
\keyword{internal}
