% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_base}
\alias{pack_uri_base}
\title{Get the base URI (directory portion) of a PackURI}
\usage{
pack_uri_base(uri)
}
\arguments{
\item{uri}{A PackURI.}
}
\value{
A character string.
}
\description{
Get the base URI (directory portion) of a PackURI
}
\keyword{internal}
