% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_ext}
\alias{pack_uri_ext}
\title{Get the file extension of a PackURI (without leading dot)}
\usage{
pack_uri_ext(uri)
}
\arguments{
\item{uri}{A PackURI.}
}
\value{
A character string.
}
\description{
Get the file extension of a PackURI (without leading dot)
}
\keyword{internal}
