% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_filename}
\alias{pack_uri_filename}
\title{Get the filename portion of a PackURI}
\usage{
pack_uri_filename(uri)
}
\arguments{
\item{uri}{A PackURI.}
}
\value{
A character string.
}
\description{
Get the filename portion of a PackURI
}
\keyword{internal}
