% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_from_rel_ref}
\alias{pack_uri_from_rel_ref}
\title{Construct a PackURI from a relative reference}
\usage{
pack_uri_from_rel_ref(base_uri, relative_ref)
}
\arguments{
\item{base_uri}{The base URI (directory) for resolving the reference.}

\item{relative_ref}{The relative reference string.}
}
\value{
A PackURI.
}
\description{
Construct a PackURI from a relative reference
}
\keyword{internal}
