% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_idx}
\alias{pack_uri_idx}
\title{Get the integer index from an array-style PackURI}
\usage{
pack_uri_idx(uri)
}
\arguments{
\item{uri}{A PackURI.}
}
\value{
An integer or NULL.
}
\description{
Returns the trailing integer from the filename, e.g. 21 for
\code{"/ppt/slides/slide21.xml"}, or NULL for singleton parts.
}
\keyword{internal}
