% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_membername}
\alias{pack_uri_membername}
\title{Get the membername (without leading slash) for ZIP storage}
\usage{
pack_uri_membername(uri)
}
\arguments{
\item{uri}{A PackURI.}
}
\value{
A character string.
}
\description{
Get the membername (without leading slash) for ZIP storage
}
\keyword{internal}
