% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_relative_ref}
\alias{pack_uri_relative_ref}
\title{Get a relative reference from a PackURI to a base URI}
\usage{
pack_uri_relative_ref(uri, base_uri)
}
\arguments{
\item{uri}{A PackURI.}

\item{base_uri}{The base URI to compute the relative reference from.}
}
\value{
A character string.
}
\description{
Get a relative reference from a PackURI to a base URI
}
\keyword{internal}
