% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{pack_uri_rels_uri}
\alias{pack_uri_rels_uri}
\title{Get the .rels URI corresponding to a PackURI}
\usage{
pack_uri_rels_uri(uri)
}
\arguments{
\item{uri}{A PackURI.}
}
\value{
A PackURI for the .rels part.
}
\description{
Get the .rels URI corresponding to a PackURI
}
\keyword{internal}
