% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{package_not_found_error}
\alias{package_not_found_error}
\title{Signal a "package not found" error}
\usage{
package_not_found_error(path)
}
\arguments{
\item{path}{The file path that was not found.}
}
\description{
Raised when a .pptx package cannot be found at the specified path.
}
\keyword{internal}
