% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{Package_open}
\alias{Package_open}
\title{Open a Package from a .pptx file}
\usage{
Package_open(pkg_file)
}
\arguments{
\item{pkg_file}{Path to a .pptx file.}
}
\value{
A Package instance.
}
\description{
Open a Package from a .pptx file
}
\keyword{internal}
