% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-serialized.R
\name{PackageReader}
\alias{PackageReader}
\title{Read an OPC package (ZIP file)}
\description{
Provides dict-like access to the binary blobs of each part in the package.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PackageReader-new}{\code{PackageReader$new()}}
\item \href{#method-PackageReader-contains}{\code{PackageReader$contains()}}
\item \href{#method-PackageReader-get_blob}{\code{PackageReader$get_blob()}}
\item \href{#method-PackageReader-rels_xml_for}{\code{PackageReader$rels_xml_for()}}
\item \href{#method-PackageReader-clone}{\code{PackageReader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageReader-new"></a>}}
\if{latex}{\out{\hypertarget{method-PackageReader-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageReader$new(pkg_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageReader-contains"></a>}}
\if{latex}{\out{\hypertarget{method-PackageReader-contains}{}}}
\subsection{Method \code{contains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageReader$contains(pack_uri)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageReader-get_blob"></a>}}
\if{latex}{\out{\hypertarget{method-PackageReader-get_blob}{}}}
\subsection{Method \code{get_blob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageReader$get_blob(pack_uri)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageReader-rels_xml_for"></a>}}
\if{latex}{\out{\hypertarget{method-PackageReader-rels_xml_for}{}}}
\subsection{Method \code{rels_xml_for()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageReader$rels_xml_for(partname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageReader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PackageReader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageReader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
