% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-serialized.R
\name{PackageWriter}
\alias{PackageWriter}
\title{Write an OPC package (ZIP file)}
\description{
Write an OPC package (ZIP file)

Write an OPC package (ZIP file)
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PackageWriter-write}{\code{PackageWriter$write()}}
\item \href{#method-PackageWriter-clone}{\code{PackageWriter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageWriter-write"></a>}}
\if{latex}{\out{\hypertarget{method-PackageWriter-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageWriter$write(pkg_file, pkg_rels, parts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PackageWriter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PackageWriter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PackageWriter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
