% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-packuri.R
\name{PackURI}
\alias{PackURI}
\title{Create a PackURI (OPC part name)}
\usage{
PackURI(uri_str)
}
\arguments{
\item{uri_str}{A string starting with "/", e.g. \code{"/ppt/slides/slide1.xml"}.}
}
\value{
A PackURI (character with S3 class).
}
\description{
A PackURI is a string that represents an absolute URI within an OPC package.
It must begin with a forward slash.
}
\keyword{internal}
