% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-text.R
\name{Paragraph}
\alias{Paragraph}
\title{Paragraph object corresponding to an \code{a:p} element.}
\description{
Paragraph object corresponding to an \code{a:p} element.

Paragraph object corresponding to an \code{a:p} element.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Paragraph-new}{\code{Paragraph$new()}}
\item \href{#method-Paragraph-element}{\code{Paragraph$element()}}
\item \href{#method-Paragraph-add_run}{\code{Paragraph$add_run()}}
\item \href{#method-Paragraph-add_line_break}{\code{Paragraph$add_line_break()}}
\item \href{#method-Paragraph-clear}{\code{Paragraph$clear()}}
\item \href{#method-Paragraph-clone}{\code{Paragraph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paragraph-new"></a>}}
\if{latex}{\out{\hypertarget{method-Paragraph-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paragraph$new(p, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paragraph-element"></a>}}
\if{latex}{\out{\hypertarget{method-Paragraph-element}{}}}
\subsection{Method \code{element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paragraph$element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paragraph-add_run"></a>}}
\if{latex}{\out{\hypertarget{method-Paragraph-add_run}{}}}
\subsection{Method \code{add_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paragraph$add_run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paragraph-add_line_break"></a>}}
\if{latex}{\out{\hypertarget{method-Paragraph-add_line_break}{}}}
\subsection{Method \code{add_line_break()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paragraph$add_line_break()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paragraph-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Paragraph-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paragraph$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paragraph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Paragraph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paragraph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
