% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{ParentedElementProxy}
\alias{ParentedElementProxy}
\title{Proxy wrapping an XML element with a parent reference}
\description{
The parent is used to resolve the \code{part} property by delegation.
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{ParentedElementProxy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParentedElementProxy-new}{\code{ParentedElementProxy$new()}}
\item \href{#method-ParentedElementProxy-clone}{\code{ParentedElementProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParentedElementProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParentedElementProxy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParentedElementProxy$new(element, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParentedElementProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParentedElementProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParentedElementProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
