% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-init.R
\name{parse_from_template}
\alias{parse_from_template}
\title{Load and parse an XML template file}
\usage{
parse_from_template(template_name)
}
\arguments{
\item{template_name}{Name of the template (without .xml extension).}
}
\value{
A wrapped element for the template root.
}
\description{
Load and parse an XML template file
}
\keyword{internal}
