% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{Part}
\alias{Part}
\title{Base class for package parts}
\description{
Provides common properties and methods for all part types.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Part-new}{\code{Part$new()}}
\item \href{#method-Part-load_rels_from_xml}{\code{Part$load_rels_from_xml()}}
\item \href{#method-Part-has_rels}{\code{Part$has_rels()}}
\item \href{#method-Part-rels_xml_bytes}{\code{Part$rels_xml_bytes()}}
\item \href{#method-Part-relate_to}{\code{Part$relate_to()}}
\item \href{#method-Part-part_related_by}{\code{Part$part_related_by()}}
\item \href{#method-Part-related_part}{\code{Part$related_part()}}
\item \href{#method-Part-target_ref}{\code{Part$target_ref()}}
\item \href{#method-Part-drop_rel}{\code{Part$drop_rel()}}
\item \href{#method-Part-clone}{\code{Part$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-new"></a>}}
\if{latex}{\out{\hypertarget{method-Part-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$new(partname, content_type, package, blob = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-load_rels_from_xml"></a>}}
\if{latex}{\out{\hypertarget{method-Part-load_rels_from_xml}{}}}
\subsection{Method \code{load_rels_from_xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$load_rels_from_xml(xml_rels, parts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-has_rels"></a>}}
\if{latex}{\out{\hypertarget{method-Part-has_rels}{}}}
\subsection{Method \code{has_rels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$has_rels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-rels_xml_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-Part-rels_xml_bytes}{}}}
\subsection{Method \code{rels_xml_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$rels_xml_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-relate_to"></a>}}
\if{latex}{\out{\hypertarget{method-Part-relate_to}{}}}
\subsection{Method \code{relate_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$relate_to(target, reltype, is_external = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-part_related_by"></a>}}
\if{latex}{\out{\hypertarget{method-Part-part_related_by}{}}}
\subsection{Method \code{part_related_by()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$part_related_by(reltype)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-related_part"></a>}}
\if{latex}{\out{\hypertarget{method-Part-related_part}{}}}
\subsection{Method \code{related_part()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$related_part(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-target_ref"></a>}}
\if{latex}{\out{\hypertarget{method-Part-target_ref}{}}}
\subsection{Method \code{target_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$target_ref(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-drop_rel"></a>}}
\if{latex}{\out{\hypertarget{method-Part-drop_rel}{}}}
\subsection{Method \code{drop_rel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$drop_rel(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Part-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Part-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Part$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
