% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{PartElementProxy}
\alias{PartElementProxy}
\title{Proxy wrapping a part's root XML element}
\description{
Used for domain objects that correspond directly to an OPC part
(e.g. Presentation wraps the root element of PresentationPart).
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{PartElementProxy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartElementProxy-new}{\code{PartElementProxy$new()}}
\item \href{#method-PartElementProxy-clone}{\code{PartElementProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartElementProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartElementProxy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartElementProxy$new(element, part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartElementProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartElementProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartElementProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
