% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{PartFactory_create}
\alias{PartFactory_create}
\title{Construct the appropriate Part subclass for a given content type}
\usage{
PartFactory_create(partname, content_type, package, blob)
}
\arguments{
\item{partname}{A PackURI.}

\item{content_type}{Content type string.}

\item{package}{The package.}

\item{blob}{Raw bytes.}
}
\value{
A Part instance.
}
\description{
Construct the appropriate Part subclass for a given content type
}
\keyword{internal}
