% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{Picture}
\alias{Picture}
\title{Shape proxy for p:pic elements}
\description{
Shape proxy for p:pic elements

Shape proxy for p:pic elements
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseShape]{rpptx::BaseShape}} -> \code{Picture}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Picture-clone}{\code{Picture$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseShape" data-id="initialize"><a href='../../rpptx/html/BaseShape.html#method-BaseShape-initialize'><code>rpptx::BaseShape$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Picture-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Picture-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Picture$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
