% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{PictureCrop}
\alias{PictureCrop}
\title{Picture crop descriptor}
\description{
Provides read/write access to the per-edge crop fractions (0.0–1.0) of a
picture. A value of 0.1 means 10\% of the edge is cropped.
Access via \code{picture$crop}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PictureCrop-new}{\code{PictureCrop$new()}}
\item \href{#method-PictureCrop-clone}{\code{PictureCrop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureCrop-new"></a>}}
\if{latex}{\out{\hypertarget{method-PictureCrop-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureCrop$new(blipFill)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureCrop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PictureCrop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureCrop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
