% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{PieSeries}
\alias{PieSeries}
\title{Pie series}
\description{
Pie series

Pie series
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:BaseSeries]{rpptx::BaseSeries}} -> \code{\link[rpptx:BaseCategorySeries]{rpptx::BaseCategorySeries}} -> \code{PieSeries}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PieSeries-clone}{\code{PieSeries$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseSeries" data-id="initialize"><a href='../../rpptx/html/BaseSeries.html#method-BaseSeries-initialize'><code>rpptx::BaseSeries$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PieSeries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PieSeries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PieSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
