% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{PlaceholderFormat}
\alias{PlaceholderFormat}
\title{Placeholder format descriptor}
\description{
Provides access to placeholder properties such as \code{idx} and \code{type}.
Accessed via \code{shape$placeholder_format} on a placeholder shape.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlaceholderFormat-new}{\code{PlaceholderFormat$new()}}
\item \href{#method-PlaceholderFormat-element}{\code{PlaceholderFormat$element()}}
\item \href{#method-PlaceholderFormat-clone}{\code{PlaceholderFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceholderFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceholderFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceholderFormat$new(ph_elm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceholderFormat-element"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceholderFormat-element}{}}}
\subsection{Method \code{element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceholderFormat$element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceholderFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceholderFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceholderFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
