% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{plot_factory}
\alias{plot_factory}
\title{Dispatch an xChart element to the appropriate plot class}
\usage{
plot_factory(xChart, chart)
}
\arguments{
\item{xChart}{A CT_xChart wrapped element.}

\item{chart}{The Chart domain object.}
}
\value{
A plot domain object.
}
\description{
Dispatch an xChart element to the appropriate plot class
}
\keyword{internal}
