% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{Point}
\alias{Point}
\title{Individual data point in a series}
\description{
Individual data point in a series

Individual data point in a series
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Point-new}{\code{Point$new()}}
\item \href{#method-Point-clone}{\code{Point$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Point-new"></a>}}
\if{latex}{\out{\hypertarget{method-Point-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Point$new(ser, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Point-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Point-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Point$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
