% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-action.R
\docType{data}
\name{PP_ACTION_TYPE}
\alias{PP_ACTION_TYPE}
\title{PP_ACTION_TYPE — type of action for a click or hover}
\format{
An object of class \code{list} of length 15.
}
\usage{
PP_ACTION_TYPE
}
\description{
Integer constants identifying the type of action associated with a mouse
click or hover on a shape or text run.
}
\keyword{datasets}
