% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-text.R
\docType{data}
\name{PP_PARAGRAPH_ALIGNMENT}
\alias{PP_PARAGRAPH_ALIGNMENT}
\title{PP_PARAGRAPH_ALIGNMENT — paragraph alignment}
\format{
An object of class \code{list} of length 8.
}
\usage{
PP_PARAGRAPH_ALIGNMENT
}
\description{
XML string values for \verb{a:pPr/@algn}. Use with paragraph \code{alignment} property.
}
\keyword{internal}
