% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-shapes.R
\docType{data}
\name{PP_PLACEHOLDER}
\alias{PP_PLACEHOLDER}
\title{PP_PLACEHOLDER — placeholder type string constants}
\format{
An object of class \code{list} of length 15.
}
\usage{
PP_PLACEHOLDER
}
\description{
String values for the \code{type} attribute of a \verb{<p:ph>} element.
}
\keyword{internal}
