% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pptx_presentation}
\alias{pptx_presentation}
\title{Create or open a PowerPoint presentation}
\usage{
pptx_presentation(pptx = NULL)
}
\arguments{
\item{pptx}{Path to a .pptx file, or NULL to use the default template.}
}
\value{
A Presentation object.
}
\description{
Create or open a PowerPoint presentation
}
