% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{Presentation}
\alias{Presentation}
\title{Presentation object}
\description{
User-facing object for a PowerPoint presentation. Wraps the
CT_Presentation XML element and delegates to PresentationPart.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:PartElementProxy]{rpptx::PartElementProxy}} -> \code{Presentation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Presentation-save}{\code{Presentation$save()}}
\item \href{#method-Presentation-clone}{\code{Presentation$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="PartElementProxy" data-id="initialize"><a href='../../rpptx/html/PartElementProxy.html#method-PartElementProxy-initialize'><code>rpptx::PartElementProxy$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Presentation-save"></a>}}
\if{latex}{\out{\hypertarget{method-Presentation-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Presentation$save(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Presentation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Presentation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Presentation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
