% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-presentation.R
\name{PresentationPart}
\alias{PresentationPart}
\title{Presentation Part}
\description{
OPC part for the presentation XML (/ppt/presentation.xml).
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:Part]{rpptx::Part}} -> \code{\link[rpptx:XmlPart]{rpptx::XmlPart}} -> \code{PresentationPart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PresentationPart-add_slide}{\code{PresentationPart$add_slide()}}
\item \href{#method-PresentationPart-get_presentation}{\code{PresentationPart$get_presentation()}}
\item \href{#method-PresentationPart-get_slide}{\code{PresentationPart$get_slide()}}
\item \href{#method-PresentationPart-related_slide}{\code{PresentationPart$related_slide()}}
\item \href{#method-PresentationPart-related_slide_master}{\code{PresentationPart$related_slide_master()}}
\item \href{#method-PresentationPart-save}{\code{PresentationPart$save()}}
\item \href{#method-PresentationPart-slide_id}{\code{PresentationPart$slide_id()}}
\item \href{#method-PresentationPart-duplicate_slide}{\code{PresentationPart$duplicate_slide()}}
\item \href{#method-PresentationPart-rename_slide_parts}{\code{PresentationPart$rename_slide_parts()}}
\item \href{#method-PresentationPart-clone}{\code{PresentationPart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="has_rels"><a href='../../rpptx/html/Part.html#method-Part-has_rels'><code>rpptx::Part$has_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="load_rels_from_xml"><a href='../../rpptx/html/Part.html#method-Part-load_rels_from_xml'><code>rpptx::Part$load_rels_from_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="part_related_by"><a href='../../rpptx/html/Part.html#method-Part-part_related_by'><code>rpptx::Part$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="relate_to"><a href='../../rpptx/html/Part.html#method-Part-relate_to'><code>rpptx::Part$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="related_part"><a href='../../rpptx/html/Part.html#method-Part-related_part'><code>rpptx::Part$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="rels_xml_bytes"><a href='../../rpptx/html/Part.html#method-Part-rels_xml_bytes'><code>rpptx::Part$rels_xml_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="target_ref"><a href='../../rpptx/html/Part.html#method-Part-target_ref'><code>rpptx::Part$target_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XmlPart" data-id="drop_rel"><a href='../../rpptx/html/XmlPart.html#method-XmlPart-drop_rel'><code>rpptx::XmlPart$drop_rel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="XmlPart" data-id="initialize"><a href='../../rpptx/html/XmlPart.html#method-XmlPart-initialize'><code>rpptx::XmlPart$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-add_slide"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-add_slide}{}}}
\subsection{Method \code{add_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$add_slide(slide_layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-get_presentation"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-get_presentation}{}}}
\subsection{Method \code{get_presentation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$get_presentation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-get_slide"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-get_slide}{}}}
\subsection{Method \code{get_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$get_slide(slide_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-related_slide"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-related_slide}{}}}
\subsection{Method \code{related_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$related_slide(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-related_slide_master"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-related_slide_master}{}}}
\subsection{Method \code{related_slide_master()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$related_slide_master(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-save"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$save(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-slide_id"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-slide_id}{}}}
\subsection{Method \code{slide_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$slide_id(slide_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-duplicate_slide"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-duplicate_slide}{}}}
\subsection{Method \code{duplicate_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$duplicate_slide(slide_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-rename_slide_parts"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-rename_slide_parts}{}}}
\subsection{Method \code{rename_slide_parts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$rename_slide_parts(rIds)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PresentationPart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PresentationPart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PresentationPart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
