% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-ns.R
\name{qn}
\alias{qn}
\title{Get the Clark-notation qualified tag name for a namespace-prefixed tag}
\usage{
qn(nsptag)
}
\arguments{
\item{nsptag}{A namespace-prefixed tag string, e.g. \code{"p:cSld"}.}
}
\value{
A string in Clark notation.
}
\description{
Converts a namespace-prefixed tag like \code{"p:cSld"} to Clark notation like
\code{"{http://schemas.openxmlformats.org/presentationml/2006/main}cSld"}.
}
\examples{
qn("p:cSld")
qn("a:r")
}
