% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-init.R
\name{register_element_cls}
\alias{register_element_cls}
\title{Register an R6 class for a given XML element tag}
\usage{
register_element_cls(nsptagname, cls)
}
\arguments{
\item{nsptagname}{Namespace-prefixed tag, e.g. \code{"p:presentation"}.}

\item{cls}{An R6ClassGenerator (e.g. the result of \code{\link[=define_oxml_element]{define_oxml_element()}}).}
}
\description{
When the XML parser encounters an element with this tag, it will be wrapped
in an instance of \code{cls} instead of the default BaseOxmlElement.
}
\keyword{internal}
