% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{register_part_type}
\alias{register_part_type}
\title{Register a Part subclass for a content type}
\usage{
register_part_type(content_type, part_cls)
}
\arguments{
\item{content_type}{Content type string.}

\item{part_cls}{An R6 class generator.}
}
\description{
Register a Part subclass for a content type
}
\keyword{internal}
