% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{Relationship}
\alias{Relationship}
\title{Value object describing a link from a part or package to another part}
\description{
Value object describing a link from a part or package to another part

Value object describing a link from a part or package to another part
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Relationship-new}{\code{Relationship$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-new"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$new(base_uri, rId, reltype, target_mode, target)}\if{html}{\out{</div>}}
}

}
}
