% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{Relationship_from_xml}
\alias{Relationship_from_xml}
\title{Create a Relationship from a CT_Relationship XML element}
\usage{
Relationship_from_xml(base_uri, rel, parts)
}
\arguments{
\item{base_uri}{Base URI string.}

\item{rel}{A CT_Relationship element.}

\item{parts}{Named list mapping partname -> Part.}
}
\value{
A Relationship object.
}
\description{
Create a Relationship from a CT_Relationship XML element
}
\keyword{internal}
