% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{Relationships}
\alias{Relationships}
\title{Collection of relationships from a part or package to other parts}
\description{
Keyed by rId. Supports dict-like access.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Relationships-new}{\code{Relationships$new()}}
\item \href{#method-Relationships-get}{\code{Relationships$get()}}
\item \href{#method-Relationships-contains}{\code{Relationships$contains()}}
\item \href{#method-Relationships-values}{\code{Relationships$values()}}
\item \href{#method-Relationships-keys}{\code{Relationships$keys()}}
\item \href{#method-Relationships-get_or_add}{\code{Relationships$get_or_add()}}
\item \href{#method-Relationships-get_or_add_ext_rel}{\code{Relationships$get_or_add_ext_rel()}}
\item \href{#method-Relationships-load_from_xml}{\code{Relationships$load_from_xml()}}
\item \href{#method-Relationships-part_with_reltype}{\code{Relationships$part_with_reltype()}}
\item \href{#method-Relationships-pop}{\code{Relationships$pop()}}
\item \href{#method-Relationships-xml_bytes}{\code{Relationships$xml_bytes()}}
\item \href{#method-Relationships-clone}{\code{Relationships$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-new"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$new(base_uri)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-get"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$get(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-contains"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-contains}{}}}
\subsection{Method \code{contains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$contains(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-values"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-values}{}}}
\subsection{Method \code{values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-keys"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-keys}{}}}
\subsection{Method \code{keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-get_or_add"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-get_or_add}{}}}
\subsection{Method \code{get_or_add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$get_or_add(reltype, target_part)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-get_or_add_ext_rel"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-get_or_add_ext_rel}{}}}
\subsection{Method \code{get_or_add_ext_rel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$get_or_add_ext_rel(reltype, target_ref)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-load_from_xml"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-load_from_xml}{}}}
\subsection{Method \code{load_from_xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$load_from_xml(base_uri, xml_rels, parts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-part_with_reltype"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-part_with_reltype}{}}}
\subsection{Method \code{part_with_reltype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$part_with_reltype(reltype)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-pop"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-pop}{}}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$pop(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-xml_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-xml_bytes}{}}}
\subsection{Method \code{xml_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$xml_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationships-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Relationships-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationships$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
