% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_functions.R
\name{replace_category_plot}
\alias{replace_category_plot}
\title{Replace category plot}
\usage{
replace_category_plot(pres, label, data)
}
\arguments{
\item{pres}{Python pptx Presentation object.}

\item{label}{(character) Label of the text box containing the text to replace.}

\item{data}{(data frame) Data frame used to replace the old table.}
}
\value{

}
\description{
Replace category plot
}
