% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_functions.R
\name{replace_donut_plot}
\alias{replace_donut_plot}
\title{Replace a donut plot}
\usage{
replace_donut_plot(pres, label, percent_favorable)
}
\arguments{
\item{pres}{Python pptx Presentation object.}

\item{label}{(character) Label of the text box containing the text to replace.}

\item{percent_favorable}{(double) Proportion of favorable responses.}
}
\value{

}
\description{
Replace a donut plot
}
