% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_functions.R
\name{replace_image}
\alias{replace_image}
\title{Replace an image in a PowerPoint text box.}
\usage{
replace_image(pres, label, new_image, new_height = TRUE)
}
\arguments{
\item{pres}{Python pptx Presentation object.}

\item{label}{(character) Label of the image to replace.}

\item{new_image}{(character) New image to replace the old image.}

\item{new_height}{(boolean) Should the aspect ratio of the new image be preserved (\code{TRUE}) or
should the image height be set to match the height of the old image (\code{FALSE}).}
}
\description{
Replace an image in a PowerPoint text box.
}
