% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_functions.R
\name{replace_table}
\alias{replace_table}
\title{Replace text in a PowerPoint table while retaining formatting.}
\usage{
replace_table(pres, label, new_table, col_names = TRUE)
}
\arguments{
\item{pres}{Python pptx Presentation object.}

\item{label}{(character) Label of the text box containing the table to replace}

\item{new_table}{(data frame) Data frame to replace elements in the old table}

\item{col_names}{(boolean) Keep new_table column names as the first row of the
new table?}
}
\description{
Replace text in a PowerPoint table while retaining formatting.
}
