% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_functions.R
\name{replace_text}
\alias{replace_text}
\title{Replace text in a PowerPoint text box while retaining formatting.}
\usage{
replace_text(pres, label, new_text)
}
\arguments{
\item{pres}{Python pptx Presentation object.}

\item{label}{(character) Label of the text box containing the text to replace.}

\item{new_text}{(character) New text to replace the old text.}
}
\value{

}
\description{
Replace text in a PowerPoint text box while retaining formatting.
}
