% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-xmlchemy.R
\name{required_attribute}
\alias{required_attribute}
\title{Define a RequiredAttribute specification}
\usage{
required_attribute(attr_name, simple_type)
}
\arguments{
\item{attr_name}{Attribute name (may be namespace-prefixed).}

\item{simple_type}{A list with \code{from_xml} and \code{to_xml} functions.}
}
\value{
A list describing the attribute spec.
}
\description{
Define a RequiredAttribute specification
}
\keyword{internal}
