% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-color.R
\name{RGBColor_from_str}
\alias{RGBColor_from_str}
\title{Create RGBColor from a 6-character hex string}
\usage{
RGBColor_from_str(hex_str)
}
\arguments{
\item{hex_str}{A 6-character hex string like "FF0000" or "#FF0000".}
}
\value{
An \code{RGBColor} object.
}
\description{
Create RGBColor from a 6-character hex string
}
