% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpptx-package.R
\docType{package}
\name{rpptx-package}
\alias{rpptx}
\alias{rpptx-package}
\title{rpptx: Create, Read, and Update PowerPoint Presentations}
\description{
Create, read, and update PowerPoint (.pptx) files. An R port of the Python 'python-pptx' library. Provides an R6-based object model for manipulating presentations, slides, shapes, text, charts, tables, and formatting properties.
}
\author{
\strong{Maintainer}: Mark Heckmann \email{heckmann.mark@gmail.com}

}
\keyword{internal}
