% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{rpptx_error}
\alias{rpptx_error}
\title{Signal an rpptx error}
\usage{
rpptx_error(message, class = NULL, call = NULL, ...)
}
\arguments{
\item{message}{Error message.}

\item{class}{Character vector of additional subclasses.}

\item{call}{The call to include in the error.}

\item{...}{Additional fields for the condition object.}
}
\value{
No return value; raises a condition.
}
\description{
Signal an rpptx error
}
\keyword{internal}
