% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-init.R
\name{rpptx_parse_xml}
\alias{rpptx_parse_xml}
\title{Parse XML bytes/string into a wrapped element}
\usage{
rpptx_parse_xml(xml)
}
\arguments{
\item{xml}{A raw vector, character string, or connection containing XML.}
}
\value{
A wrapped BaseOxmlElement (or registered subclass) for the root.
}
\description{
Parse XML bytes/string into a wrapped element
}
\keyword{internal}
