% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-constants.R
\docType{data}
\name{RT}
\alias{RT}
\title{Open XML relationship type URIs.}
\format{
An object of class \code{list} of length 75.
}
\usage{
RT
}
\description{
Open XML relationship type URIs.
}
\keyword{internal}
