% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-constants.R
\docType{data}
\name{RTM}
\alias{RTM}
\title{Open XML relationship target modes.}
\format{
An object of class \code{list} of length 2.
}
\usage{
RTM
}
\description{
Open XML relationship target modes.
}
\keyword{internal}
