% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-text.R
\name{Run}
\alias{Run}
\title{Text run object corresponding to an \code{a:r} element.}
\description{
Text run object corresponding to an \code{a:r} element.

Text run object corresponding to an \code{a:r} element.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Run-new}{\code{Run$new()}}
\item \href{#method-Run-element}{\code{Run$element()}}
\item \href{#method-Run-clone}{\code{Run$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-new"></a>}}
\if{latex}{\out{\hypertarget{method-Run-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$new(r, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-element"></a>}}
\if{latex}{\out{\hypertarget{method-Run-element}{}}}
\subsection{Method \code{element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Run-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
