% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_functions.R
\name{save_pres}
\alias{save_pres}
\title{Save presentation as a new file}
\usage{
save_pres(pres, path)
}
\arguments{
\item{pres}{(Presentation object) Python pptx Presentation object to save.}

\item{path}{(character) Path indicating where to save the presentation.}
}
\description{
Save presentation as a new file
}
