% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-oxml.R
\name{serialize_part_xml}
\alias{serialize_part_xml}
\title{Serialize a BaseOxmlElement to XML bytes}
\usage{
serialize_part_xml(element)
}
\arguments{
\item{element}{A BaseOxmlElement.}
}
\value{
A raw vector of XML bytes.
}
\description{
Serialize a BaseOxmlElement to XML bytes
}
\keyword{internal}
