% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{series_factory}
\alias{series_factory}
\title{Dispatch a c:ser element to the appropriate series class}
\usage{
series_factory(ser)
}
\arguments{
\item{ser}{A CT_SeriesComposite wrapped element.}
}
\value{
A series domain object.
}
\description{
Dispatch a c:ser element to the appropriate series class
}
\keyword{internal}
