% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{SeriesCollection}
\alias{SeriesCollection}
\title{Collection of series in a chart or plot}
\description{
Collection of series in a chart or plot

Collection of series in a chart or plot
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeriesCollection-new}{\code{SeriesCollection$new()}}
\item \href{#method-SeriesCollection-get}{\code{SeriesCollection$get()}}
\item \href{#method-SeriesCollection-to_list}{\code{SeriesCollection$to_list()}}
\item \href{#method-SeriesCollection-clone}{\code{SeriesCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesCollection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesCollection$new(parent_elm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesCollection-get"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesCollection-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesCollection$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesCollection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesCollection-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesCollection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
