% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml-effect.R
\name{ShadowFormat}
\alias{ShadowFormat}
\title{Shape shadow effect}
\description{
Provides access to the shadow effect for a shape. Obtain via \code{shape$shadow}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShadowFormat-new}{\code{ShadowFormat$new()}}
\item \href{#method-ShadowFormat-clone}{\code{ShadowFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShadowFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShadowFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShadowFormat$new(spPr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShadowFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShadowFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShadowFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
