% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-base.R
\name{Shape}
\alias{Shape}
\title{Shape proxy for p:sp elements}
\description{
Covers autoshapes, text boxes, and placeholders.
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseShape]{rpptx::BaseShape}} -> \code{Shape}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Shape-clone}{\code{Shape$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseShape" data-id="initialize"><a href='../../rpptx/html/BaseShape.html#method-BaseShape-initialize'><code>rpptx::BaseShape$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shape-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Shape-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
