% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{ShapeHyperlink}
\alias{ShapeHyperlink}
\title{Hyperlink on a shape}
\description{
Provides read/write access to a URL hyperlink on a shape.
Access via \code{shape$click_action$hyperlink}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShapeHyperlink-new}{\code{ShapeHyperlink$new()}}
\item \href{#method-ShapeHyperlink-clone}{\code{ShapeHyperlink$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapeHyperlink-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShapeHyperlink-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapeHyperlink$new(cNvPr, slide_part, hover = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapeHyperlink-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShapeHyperlink-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapeHyperlink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
