% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-simpletypes.R
\name{simple_type}
\alias{simple_type}
\title{Create a simple type converter}
\usage{
simple_type(from_xml, to_xml)
}
\arguments{
\item{from_xml}{Function(str) -> R value.}

\item{to_xml}{Function(R value) -> str.}
}
\value{
A list with \code{from_xml} and \code{to_xml} elements.
}
\description{
Create a simple type converter
}
\keyword{internal}
