% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{Slide}
\alias{Slide}
\title{Slide object}
\description{
Provides access to slide properties.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:PartElementProxy]{rpptx::PartElementProxy}} -> \code{\link[rpptx:.BaseSlide]{rpptx::.BaseSlide}} -> \code{Slide}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Slide-clone}{\code{Slide$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="PartElementProxy" data-id="initialize"><a href='../../rpptx/html/PartElementProxy.html#method-PartElementProxy-initialize'><code>rpptx::PartElementProxy$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slide-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Slide-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slide$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
