% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{SlideLayout}
\alias{SlideLayout}
\title{Slide layout object}
\description{
Slide layout object

Slide layout object
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:PartElementProxy]{rpptx::PartElementProxy}} -> \code{\link[rpptx:.BaseSlide]{rpptx::.BaseSlide}} -> \code{SlideLayout}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlideLayout-iter_cloneable_placeholders}{\code{SlideLayout$iter_cloneable_placeholders()}}
\item \href{#method-SlideLayout-clone}{\code{SlideLayout$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="PartElementProxy" data-id="initialize"><a href='../../rpptx/html/PartElementProxy.html#method-PartElementProxy-initialize'><code>rpptx::PartElementProxy$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayout-iter_cloneable_placeholders"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayout-iter_cloneable_placeholders}{}}}
\subsection{Method \code{iter_cloneable_placeholders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayout$iter_cloneable_placeholders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayout-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayout-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
