% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{SlideLayouts}
\alias{SlideLayouts}
\title{Sequence of slide layouts for a slide master}
\description{
Sequence of slide layouts for a slide master

Sequence of slide layouts for a slide master
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{SlideLayouts}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlideLayouts-new}{\code{SlideLayouts$new()}}
\item \href{#method-SlideLayouts-get}{\code{SlideLayouts$get()}}
\item \href{#method-SlideLayouts-get_by_name}{\code{SlideLayouts$get_by_name()}}
\item \href{#method-SlideLayouts-index}{\code{SlideLayouts$index()}}
\item \href{#method-SlideLayouts-to_list}{\code{SlideLayouts$to_list()}}
\item \href{#method-SlideLayouts-remove}{\code{SlideLayouts$remove()}}
\item \href{#method-SlideLayouts-clone}{\code{SlideLayouts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-new"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$new(sldLayoutIdLst, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-get"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-get_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-get_by_name}{}}}
\subsection{Method \code{get_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$get_by_name(name, default = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-index"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-index}{}}}
\subsection{Method \code{index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$index(slide_layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-remove"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$remove(slide_layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideLayouts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlideLayouts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideLayouts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
