% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{SlideMasterPlaceholders}
\alias{SlideMasterPlaceholders}
\title{Placeholder collection for a slide master}
\description{
Like SlidePlaceholders but \code{get()} accepts a placeholder type string
(e.g. "title", "body") rather than an integer idx.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{\link[rpptx:SlidePlaceholders]{rpptx::SlidePlaceholders}} -> \code{SlideMasterPlaceholders}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlideMasterPlaceholders-get}{\code{SlideMasterPlaceholders$get()}}
\item \href{#method-SlideMasterPlaceholders-clone}{\code{SlideMasterPlaceholders$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="SlidePlaceholders" data-id="get_at"><a href='../../rpptx/html/SlidePlaceholders.html#method-SlidePlaceholders-get_at'><code>rpptx::SlidePlaceholders$get_at()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="SlidePlaceholders" data-id="initialize"><a href='../../rpptx/html/SlidePlaceholders.html#method-SlidePlaceholders-initialize'><code>rpptx::SlidePlaceholders$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="SlidePlaceholders" data-id="to_list"><a href='../../rpptx/html/SlidePlaceholders.html#method-SlidePlaceholders-to_list'><code>rpptx::SlidePlaceholders$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideMasterPlaceholders-get"></a>}}
\if{latex}{\out{\hypertarget{method-SlideMasterPlaceholders-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideMasterPlaceholders$get(ph_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideMasterPlaceholders-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlideMasterPlaceholders-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideMasterPlaceholders$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
