% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{SlideMasters}
\alias{SlideMasters}
\title{Sequence of slide masters in a presentation}
\description{
Sequence of slide masters in a presentation

Sequence of slide masters in a presentation
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{SlideMasters}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlideMasters-new}{\code{SlideMasters$new()}}
\item \href{#method-SlideMasters-get}{\code{SlideMasters$get()}}
\item \href{#method-SlideMasters-to_list}{\code{SlideMasters$to_list()}}
\item \href{#method-SlideMasters-clone}{\code{SlideMasters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideMasters-new"></a>}}
\if{latex}{\out{\hypertarget{method-SlideMasters-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideMasters$new(sldMasterIdLst, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideMasters-get"></a>}}
\if{latex}{\out{\hypertarget{method-SlideMasters-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideMasters$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideMasters-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SlideMasters-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideMasters$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideMasters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlideMasters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideMasters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
