% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{SlidePlaceholders}
\alias{SlidePlaceholders}
\title{Placeholder collection for a slide}
\description{
Provides access to placeholder shapes on a slide. Supports indexed access
by placeholder \code{idx} via \code{[[}, \code{length()}, and \code{to_list()}.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{SlidePlaceholders}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlidePlaceholders-new}{\code{SlidePlaceholders$new()}}
\item \href{#method-SlidePlaceholders-get}{\code{SlidePlaceholders$get()}}
\item \href{#method-SlidePlaceholders-get_at}{\code{SlidePlaceholders$get_at()}}
\item \href{#method-SlidePlaceholders-to_list}{\code{SlidePlaceholders$to_list()}}
\item \href{#method-SlidePlaceholders-clone}{\code{SlidePlaceholders$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlidePlaceholders-new"></a>}}
\if{latex}{\out{\hypertarget{method-SlidePlaceholders-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlidePlaceholders$new(spTree, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlidePlaceholders-get"></a>}}
\if{latex}{\out{\hypertarget{method-SlidePlaceholders-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlidePlaceholders$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlidePlaceholders-get_at"></a>}}
\if{latex}{\out{\hypertarget{method-SlidePlaceholders-get_at}{}}}
\subsection{Method \code{get_at()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlidePlaceholders$get_at(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlidePlaceholders-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SlidePlaceholders-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlidePlaceholders$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlidePlaceholders-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlidePlaceholders-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlidePlaceholders$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
