% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{Slides}
\alias{Slides}
\title{Sequence of slides in a presentation}
\description{
Supports indexed access (1-based), \code{length()}, and iteration via \code{to_list()}.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{Slides}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Slides-new}{\code{Slides$new()}}
\item \href{#method-Slides-get}{\code{Slides$get()}}
\item \href{#method-Slides-add_slide}{\code{Slides$add_slide()}}
\item \href{#method-Slides-get_by_id}{\code{Slides$get_by_id()}}
\item \href{#method-Slides-index}{\code{Slides$index()}}
\item \href{#method-Slides-to_list}{\code{Slides$to_list()}}
\item \href{#method-Slides-duplicate_slide}{\code{Slides$duplicate_slide()}}
\item \href{#method-Slides-delete}{\code{Slides$delete()}}
\item \href{#method-Slides-move}{\code{Slides$move()}}
\item \href{#method-Slides-clone}{\code{Slides$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-new"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$new(sldIdLst, prs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-get"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-add_slide"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-add_slide}{}}}
\subsection{Method \code{add_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$add_slide(slide_layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-get_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-get_by_id}{}}}
\subsection{Method \code{get_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$get_by_id(slide_id, default = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-index"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-index}{}}}
\subsection{Method \code{index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$index(slide)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-duplicate_slide"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-duplicate_slide}{}}}
\subsection{Method \code{duplicate_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$duplicate_slide(slide)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$delete(slide)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-move"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-move}{}}}
\subsection{Method \code{move()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$move(slide, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Slides-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Slides-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Slides$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
