% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{SlideShapes}
\alias{SlideShapes}
\title{Shape collection for a slide}
\description{
Supports indexed access (1-based), \code{length()}, and iteration via \code{to_list()}.
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:ElementProxy]{rpptx::ElementProxy}} -> \code{\link[rpptx:ParentedElementProxy]{rpptx::ParentedElementProxy}} -> \code{SlideShapes}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlideShapes-new}{\code{SlideShapes$new()}}
\item \href{#method-SlideShapes-get}{\code{SlideShapes$get()}}
\item \href{#method-SlideShapes-to_list}{\code{SlideShapes$to_list()}}
\item \href{#method-SlideShapes-add_textbox}{\code{SlideShapes$add_textbox()}}
\item \href{#method-SlideShapes-add_shape}{\code{SlideShapes$add_shape()}}
\item \href{#method-SlideShapes-add_table}{\code{SlideShapes$add_table()}}
\item \href{#method-SlideShapes-clone_layout_placeholders}{\code{SlideShapes$clone_layout_placeholders()}}
\item \href{#method-SlideShapes-clone}{\code{SlideShapes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-new"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$new(spTree, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-get"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-add_textbox"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-add_textbox}{}}}
\subsection{Method \code{add_textbox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$add_textbox(left, top, width, height)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-add_shape"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-add_shape}{}}}
\subsection{Method \code{add_shape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$add_shape(auto_shape_type, left, top, width, height)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-add_table"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-add_table}{}}}
\subsection{Method \code{add_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$add_table(rows, cols, left, top, width, height)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-clone_layout_placeholders"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-clone_layout_placeholders}{}}}
\subsection{Method \code{clone_layout_placeholders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$clone_layout_placeholders(slide_layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlideShapes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlideShapes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlideShapes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
