% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{Table}
\alias{Table}
\title{Table proxy}
\description{
Provides access to table cells, rows, columns, and formatting flags.
Access via \code{graphic_frame$table}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Table-new}{\code{Table$new()}}
\item \href{#method-Table-cell}{\code{Table$cell()}}
\item \href{#method-Table-iter_cells}{\code{Table$iter_cells()}}
\item \href{#method-Table-clone}{\code{Table$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-new"></a>}}
\if{latex}{\out{\hypertarget{method-Table-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$new(tbl_elm, graphic_frame)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-cell"></a>}}
\if{latex}{\out{\hypertarget{method-Table-cell}{}}}
\subsection{Method \code{cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$cell(row_idx, col_idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-iter_cells"></a>}}
\if{latex}{\out{\hypertarget{method-Table-iter_cells}{}}}
\subsection{Method \code{iter_cells()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$iter_cells()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Table-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
