% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{TableCell}
\alias{TableCell}
\title{Table cell proxy}
\description{
Provides access to cell text, fill, margins, and merge state.
Access via \code{table$cell(row, col)} or \code{table$rows[[i]]$cells[[j]]}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableCell-new}{\code{TableCell$new()}}
\item \href{#method-TableCell-merge}{\code{TableCell$merge()}}
\item \href{#method-TableCell-clone}{\code{TableCell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCell-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableCell-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCell$new(tc, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCell-merge"></a>}}
\if{latex}{\out{\hypertarget{method-TableCell-merge}{}}}
\subsection{Method \code{merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCell$merge(other_cell)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCell-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableCell-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
