% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{TableCells}
\alias{TableCells}
\title{Cell collection for a table row}
\description{
Cell collection for a table row

Cell collection for a table row
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableCells-new}{\code{TableCells$new()}}
\item \href{#method-TableCells-get}{\code{TableCells$get()}}
\item \href{#method-TableCells-to_list}{\code{TableCells$to_list()}}
\item \href{#method-TableCells-clone}{\code{TableCells$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCells-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableCells-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCells$new(tr, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCells-get"></a>}}
\if{latex}{\out{\hypertarget{method-TableCells-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCells$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCells-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TableCells-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCells$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableCells-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableCells-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableCells$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
