% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{TableColumn}
\alias{TableColumn}
\title{Table column proxy}
\description{
Provides access to column width.
Access via \code{table$columns[[i]]}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableColumn-new}{\code{TableColumn$new()}}
\item \href{#method-TableColumn-clone}{\code{TableColumn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableColumn-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableColumn-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableColumn$new(gridCol, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableColumn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableColumn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableColumn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
