% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{TableColumns}
\alias{TableColumns}
\title{Column collection for a table}
\description{
Column collection for a table

Column collection for a table
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableColumns-new}{\code{TableColumns$new()}}
\item \href{#method-TableColumns-get}{\code{TableColumns$get()}}
\item \href{#method-TableColumns-to_list}{\code{TableColumns$to_list()}}
\item \href{#method-TableColumns-clone}{\code{TableColumns$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableColumns-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableColumns-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableColumns$new(tbl, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableColumns-get"></a>}}
\if{latex}{\out{\hypertarget{method-TableColumns-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableColumns$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableColumns-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TableColumns-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableColumns$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableColumns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableColumns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableColumns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
