% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{TableRow}
\alias{TableRow}
\title{Table row proxy}
\description{
Provides access to row height and cells.
Access via \code{table$rows[[i]]}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableRow-new}{\code{TableRow$new()}}
\item \href{#method-TableRow-clone}{\code{TableRow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableRow-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableRow-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableRow$new(tr, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableRow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableRow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableRow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
