% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{TableRows}
\alias{TableRows}
\title{Row collection for a table}
\description{
Row collection for a table

Row collection for a table
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableRows-new}{\code{TableRows$new()}}
\item \href{#method-TableRows-get}{\code{TableRows$get()}}
\item \href{#method-TableRows-to_list}{\code{TableRows$to_list()}}
\item \href{#method-TableRows-clone}{\code{TableRows$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableRows-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableRows-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableRows$new(tbl, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableRows-get"></a>}}
\if{latex}{\out{\hypertarget{method-TableRows-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableRows$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableRows-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TableRows-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableRows$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableRows-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableRows-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableRows$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
