% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-text.R
\name{TextFrame}
\alias{TextFrame}
\title{Text frame corresponding to a \code{p:txBody} element.}
\description{
Contains the text of a shape. Access via \code{shape$text_frame}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextFrame-new}{\code{TextFrame$new()}}
\item \href{#method-TextFrame-element}{\code{TextFrame$element()}}
\item \href{#method-TextFrame-add_paragraph}{\code{TextFrame$add_paragraph()}}
\item \href{#method-TextFrame-clear}{\code{TextFrame$clear()}}
\item \href{#method-TextFrame-clone}{\code{TextFrame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFrame-new"></a>}}
\if{latex}{\out{\hypertarget{method-TextFrame-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFrame$new(txBody, parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFrame-element"></a>}}
\if{latex}{\out{\hypertarget{method-TextFrame-element}{}}}
\subsection{Method \code{element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFrame$element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFrame-add_paragraph"></a>}}
\if{latex}{\out{\hypertarget{method-TextFrame-add_paragraph}{}}}
\subsection{Method \code{add_paragraph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFrame$add_paragraph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFrame-clear"></a>}}
\if{latex}{\out{\hypertarget{method-TextFrame-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFrame$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
