% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{TickLabels}
\alias{TickLabels}
\title{Tick label formatting object for a chart axis}
\description{
Tick label formatting object for a chart axis

Tick label formatting object for a chart axis
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TickLabels-new}{\code{TickLabels$new()}}
\item \href{#method-TickLabels-clone}{\code{TickLabels$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TickLabels-new"></a>}}
\if{latex}{\out{\hypertarget{method-TickLabels-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TickLabels$new(xAx_elm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TickLabels-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TickLabels-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TickLabels$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
