% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{ValueAxis}
\alias{ValueAxis}
\title{Value axis domain object}
\description{
Value axis domain object

Value axis domain object
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseAxis]{rpptx::BaseAxis}} -> \code{ValueAxis}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValueAxis-clone}{\code{ValueAxis$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseAxis" data-id="initialize"><a href='../../rpptx/html/BaseAxis.html#method-BaseAxis-initialize'><code>rpptx::BaseAxis$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueAxis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValueAxis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueAxis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
