% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{Video}
\alias{Video}
\title{Video value object}
\description{
Wraps a video bytestream and its MIME type. Used by \code{SlideShapes$add_movie()}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Video-new}{\code{Video$new()}}
\item \href{#method-Video-clone}{\code{Video$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-new"></a>}}
\if{latex}{\out{\hypertarget{method-Video-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$new(blob, mime_type = NULL, filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Video-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
