% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-init.R
\name{wrap_element}
\alias{wrap_element}
\title{Wrap an xml2 node in the appropriate R6 element class}
\usage{
wrap_element(node)
}
\arguments{
\item{node}{An xml2 xml_node.}
}
\value{
A BaseOxmlElement (or subclass) instance.
}
\description{
Looks up the node's tag in the element registry and returns an instance
of the registered class, or a BaseOxmlElement if no class is registered.
}
\keyword{internal}
