% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_AXIS_CROSSES}
\alias{XL_AXIS_CROSSES}
\title{XL_AXIS_CROSSES — where the other axis crosses this axis}
\format{
An object of class \code{list} of length 4.
}
\usage{
XL_AXIS_CROSSES
}
\description{
XML string values for \verb{c:crosses/@val}. Use with axis \code{crosses} property.
}
\keyword{datasets}
