% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_CATEGORY_TYPE}
\alias{XL_CATEGORY_TYPE}
\title{XL_CATEGORY_TYPE — category axis scale type}
\format{
An object of class \code{list} of length 3.
}
\usage{
XL_CATEGORY_TYPE
}
\description{
Integer constants (read-only identifiers, not XML-encoded directly).
}
\keyword{datasets}
