% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_CHART_TYPE}
\alias{XL_CHART_TYPE}
\title{XL_CHART_TYPE — chart type enumeration}
\format{
An object of class \code{list} of length 73.
}
\usage{
XL_CHART_TYPE
}
\description{
Integer constants specifying chart types, mirroring XlChartType.
}
\keyword{datasets}
