% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_DATA_LABEL_POSITION}
\alias{XL_DATA_LABEL_POSITION}
\title{XL_DATA_LABEL_POSITION — position of a data label}
\format{
An object of class \code{list} of length 10.
}
\usage{
XL_DATA_LABEL_POSITION
}
\description{
XML string values for \verb{c:dLblPos/@val}.
}
\keyword{internal}
