% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_LABEL_POSITION}
\alias{XL_LABEL_POSITION}
\title{XL_LABEL_POSITION — alias for XL_DATA_LABEL_POSITION}
\format{
An object of class \code{list} of length 10.
}
\usage{
XL_LABEL_POSITION
}
\description{
XL_LABEL_POSITION — alias for XL_DATA_LABEL_POSITION
}
\keyword{internal}
