% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_LEGEND_POSITION}
\alias{XL_LEGEND_POSITION}
\title{XL_LEGEND_POSITION — position of chart legend}
\format{
An object of class \code{list} of length 6.
}
\usage{
XL_LEGEND_POSITION
}
\description{
XML string values for \verb{c:legendPos/@val}.
}
\keyword{datasets}
