% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_MARKER_STYLE}
\alias{XL_MARKER_STYLE}
\title{XL_MARKER_STYLE — shape of a data point marker}
\format{
An object of class \code{list} of length 12.
}
\usage{
XL_MARKER_STYLE
}
\description{
XML string values for \verb{c:symbol/@val}.
}
\keyword{datasets}
