% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_TICK_LABEL_POSITION}
\alias{XL_TICK_LABEL_POSITION}
\title{XL_TICK_LABEL_POSITION — position of tick-mark labels on a chart axis}
\format{
An object of class \code{list} of length 4.
}
\usage{
XL_TICK_LABEL_POSITION
}
\description{
XML string values for \verb{c:tickLblPos/@val}.
}
\keyword{internal}
