% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-chart.R
\docType{data}
\name{XL_TICK_MARK}
\alias{XL_TICK_MARK}
\title{XL_TICK_MARK — type of axis tick mark}
\format{
An object of class \code{list} of length 4.
}
\usage{
XL_TICK_MARK
}
\description{
XML string values for \verb{c:majorTickMark/@val} and \verb{c:minorTickMark/@val}.
}
\keyword{internal}
