% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{XmlPart}
\alias{XmlPart}
\title{Base class for parts containing an XML payload}
\description{
Provides additional methods for parsing/reserializing XML and managing
relationships. Most package parts are XmlParts.
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:Part]{rpptx::Part}} -> \code{XmlPart}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XmlPart-new}{\code{XmlPart$new()}}
\item \href{#method-XmlPart-drop_rel}{\code{XmlPart$drop_rel()}}
\item \href{#method-XmlPart-clone}{\code{XmlPart$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="has_rels"><a href='../../rpptx/html/Part.html#method-Part-has_rels'><code>rpptx::Part$has_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="load_rels_from_xml"><a href='../../rpptx/html/Part.html#method-Part-load_rels_from_xml'><code>rpptx::Part$load_rels_from_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="part_related_by"><a href='../../rpptx/html/Part.html#method-Part-part_related_by'><code>rpptx::Part$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="relate_to"><a href='../../rpptx/html/Part.html#method-Part-relate_to'><code>rpptx::Part$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="related_part"><a href='../../rpptx/html/Part.html#method-Part-related_part'><code>rpptx::Part$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="rels_xml_bytes"><a href='../../rpptx/html/Part.html#method-Part-rels_xml_bytes'><code>rpptx::Part$rels_xml_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="Part" data-id="target_ref"><a href='../../rpptx/html/Part.html#method-Part-target_ref'><code>rpptx::Part$target_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XmlPart-new"></a>}}
\if{latex}{\out{\hypertarget{method-XmlPart-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XmlPart$new(partname, content_type, package, element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XmlPart-drop_rel"></a>}}
\if{latex}{\out{\hypertarget{method-XmlPart-drop_rel}{}}}
\subsection{Method \code{drop_rel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XmlPart$drop_rel(rId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XmlPart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XmlPart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XmlPart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
