% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc-package.R
\name{XmlPart_load}
\alias{XmlPart_load}
\title{Load an XmlPart from a blob}
\usage{
XmlPart_load(cls, partname, content_type, package, blob)
}
\arguments{
\item{cls}{The R6 class generator to use.}

\item{partname}{A PackURI.}

\item{content_type}{Content type string.}

\item{package}{The package.}

\item{blob}{Raw bytes of XML.}
}
\value{
An XmlPart instance.
}
\description{
Load an XmlPart from a blob
}
\keyword{internal}
