% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{XyPoints}
\alias{XyPoints}
\title{Point collection for XY (scatter) series}
\description{
Point collection for XY (scatter) series

Point collection for XY (scatter) series
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XyPoints-new}{\code{XyPoints$new()}}
\item \href{#method-XyPoints-get}{\code{XyPoints$get()}}
\item \href{#method-XyPoints-clone}{\code{XyPoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyPoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-XyPoints-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyPoints$new(ser)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyPoints-get"></a>}}
\if{latex}{\out{\hypertarget{method-XyPoints-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyPoints$get(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyPoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XyPoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyPoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
