% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-chart.R
\name{XySeries}
\alias{XySeries}
\title{XY scatter series}
\description{
XY scatter series

XY scatter series
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:BaseSeries]{rpptx::BaseSeries}} -> \code{XySeries}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XySeries-clone}{\code{XySeries$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseSeries" data-id="initialize"><a href='../../rpptx/html/BaseSeries.html#method-BaseSeries-initialize'><code>rpptx::BaseSeries$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XySeries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XySeries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XySeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
