% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-xlsx.R
\name{XyWorkbookWriter}
\alias{XyWorkbookWriter}
\title{Workbook writer for XY (scatter) charts}
\description{
Workbook writer for XY (scatter) charts

Workbook writer for XY (scatter) charts
}
\keyword{internal}
\section{Super class}{
\code{rpptx::.BaseWorkbookWriter} -> \code{XyWorkbookWriter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XyWorkbookWriter-series_name_ref}{\code{XyWorkbookWriter$series_name_ref()}}
\item \href{#method-XyWorkbookWriter-x_values_ref}{\code{XyWorkbookWriter$x_values_ref()}}
\item \href{#method-XyWorkbookWriter-y_values_ref}{\code{XyWorkbookWriter$y_values_ref()}}
\item \href{#method-XyWorkbookWriter-clone}{\code{XyWorkbookWriter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic=".BaseWorkbookWriter" data-id="initialize"><a href='../../rpptx/html/.BaseWorkbookWriter.html#method-.BaseWorkbookWriter-initialize'><code>rpptx::.BaseWorkbookWriter$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyWorkbookWriter-series_name_ref"></a>}}
\if{latex}{\out{\hypertarget{method-XyWorkbookWriter-series_name_ref}{}}}
\subsection{Method \code{series_name_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyWorkbookWriter$series_name_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyWorkbookWriter-x_values_ref"></a>}}
\if{latex}{\out{\hypertarget{method-XyWorkbookWriter-x_values_ref}{}}}
\subsection{Method \code{x_values_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyWorkbookWriter$x_values_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyWorkbookWriter-y_values_ref"></a>}}
\if{latex}{\out{\hypertarget{method-XyWorkbookWriter-y_values_ref}{}}}
\subsection{Method \code{y_values_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyWorkbookWriter$y_values_ref(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XyWorkbookWriter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XyWorkbookWriter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XyWorkbookWriter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
