% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-xmlchemy.R
\name{zero_or_more}
\alias{zero_or_more}
\title{Define a ZeroOrMore child element specification}
\usage{
zero_or_more(nsptagname, successors = character(0))
}
\arguments{
\item{nsptagname}{Namespace-prefixed tag, e.g. \code{"p:sldIdLst"}.}

\item{successors}{Character vector of successor tag names for ordering.}
}
\value{
A list describing the child element spec.
}
\description{
Define a ZeroOrMore child element specification
}
\keyword{internal}
