% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxml-shapes.R
\name{CT_GroupShape}
\alias{CT_GroupShape}
\title{Custom element class for p:spTree and p:grpSp elements}
\description{
Custom element class for p:spTree and p:grpSp elements

Custom element class for p:spTree and p:grpSp elements
}
\keyword{internal}
\section{Super classes}{
\code{\link[rpptx:BaseOxmlElement]{rpptx::BaseOxmlElement}} -> \code{\link[rpptx:BaseShapeElement]{rpptx::BaseShapeElement}} -> \code{CT_GroupShape}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CT_GroupShape-iter_shape_elms}{\code{CT_GroupShape$iter_shape_elms()}}
\item \href{#method-CT_GroupShape-iter_ph_elms}{\code{CT_GroupShape$iter_ph_elms()}}
\item \href{#method-CT_GroupShape-max_shape_id}{\code{CT_GroupShape$max_shape_id()}}
\item \href{#method-CT_GroupShape-next_shape_id}{\code{CT_GroupShape$next_shape_id()}}
\item \href{#method-CT_GroupShape-add_textbox}{\code{CT_GroupShape$add_textbox()}}
\item \href{#method-CT_GroupShape-add_autoshape}{\code{CT_GroupShape$add_autoshape()}}
\item \href{#method-CT_GroupShape-add_placeholder}{\code{CT_GroupShape$add_placeholder()}}
\item \href{#method-CT_GroupShape-add_table}{\code{CT_GroupShape$add_table()}}
\item \href{#method-CT_GroupShape-add_chart}{\code{CT_GroupShape$add_chart()}}
\item \href{#method-CT_GroupShape-add_pic}{\code{CT_GroupShape$add_pic()}}
\item \href{#method-CT_GroupShape-add_cxnSp}{\code{CT_GroupShape$add_cxnSp()}}
\item \href{#method-CT_GroupShape-add_grpSp}{\code{CT_GroupShape$add_grpSp()}}
\item \href{#method-CT_GroupShape-clone}{\code{CT_GroupShape$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="append_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-append_child'><code>rpptx::BaseOxmlElement$append_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="find"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-find'><code>rpptx::BaseOxmlElement$find()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="findall"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-findall'><code>rpptx::BaseOxmlElement$findall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="first_child_found_in"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-first_child_found_in'><code>rpptx::BaseOxmlElement$first_child_found_in()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attr'><code>rpptx::BaseOxmlElement$get_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_attrs"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_attrs'><code>rpptx::BaseOxmlElement$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="get_node"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-get_node'><code>rpptx::BaseOxmlElement$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="initialize"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-initialize'><code>rpptx::BaseOxmlElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="insert_element_before"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-insert_element_before'><code>rpptx::BaseOxmlElement$insert_element_before()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_all"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_all'><code>rpptx::BaseOxmlElement$remove_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_attr'><code>rpptx::BaseOxmlElement$remove_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="remove_child"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-remove_child'><code>rpptx::BaseOxmlElement$remove_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="set_attr"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-set_attr'><code>rpptx::BaseOxmlElement$set_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="to_xml"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-to_xml'><code>rpptx::BaseOxmlElement$to_xml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseOxmlElement" data-id="xpath"><a href='../../rpptx/html/BaseOxmlElement.html#method-BaseOxmlElement-xpath'><code>rpptx::BaseOxmlElement$xpath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseShapeElement" data-id="get_or_add_ln"><a href='../../rpptx/html/BaseShapeElement.html#method-BaseShapeElement-get_or_add_ln'><code>rpptx::BaseShapeElement$get_or_add_ln()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="BaseShapeElement" data-id="get_or_add_xfrm"><a href='../../rpptx/html/BaseShapeElement.html#method-BaseShapeElement-get_or_add_xfrm'><code>rpptx::BaseShapeElement$get_or_add_xfrm()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-iter_shape_elms"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-iter_shape_elms}{}}}
\subsection{Method \code{iter_shape_elms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$iter_shape_elms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-iter_ph_elms"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-iter_ph_elms}{}}}
\subsection{Method \code{iter_ph_elms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$iter_ph_elms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-max_shape_id"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-max_shape_id}{}}}
\subsection{Method \code{max_shape_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$max_shape_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-next_shape_id"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-next_shape_id}{}}}
\subsection{Method \code{next_shape_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$next_shape_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_textbox"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_textbox}{}}}
\subsection{Method \code{add_textbox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_textbox(id, name, x, y, cx, cy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_autoshape"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_autoshape}{}}}
\subsection{Method \code{add_autoshape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_autoshape(id, name, prst, x, y, cx, cy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_placeholder}{}}}
\subsection{Method \code{add_placeholder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_placeholder(id, name, ph_type, orient, sz, idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_table"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_table}{}}}
\subsection{Method \code{add_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_table(id, name, rows, cols, x, y, cx, cy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_chart"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_chart}{}}}
\subsection{Method \code{add_chart()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_chart(id, name, rId, x, y, cx, cy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_pic"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_pic}{}}}
\subsection{Method \code{add_pic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_pic(id, name, desc, rId, x, y, cx, cy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_cxnSp"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_cxnSp}{}}}
\subsection{Method \code{add_cxnSp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_cxnSp(
  id,
  name,
  prst,
  x,
  y,
  cx,
  cy,
  flipH = FALSE,
  flipV = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-add_grpSp"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-add_grpSp}{}}}
\subsection{Method \code{add_grpSp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$add_grpSp(id, name, shape_elms)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CT_GroupShape-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CT_GroupShape-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CT_GroupShape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
