% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-shapes.R
\docType{data}
\name{MSO_CONNECTOR_TYPE}
\alias{MSO_CONNECTOR_TYPE}
\alias{MSO_CONNECTOR}
\title{Connector type constants (MSO_CONNECTOR_TYPE)}
\format{
An object of class \code{list} of length 4.

An object of class \code{list} of length 4.
}
\usage{
MSO_CONNECTOR_TYPE

MSO_CONNECTOR
}
\description{
Members have \verb{$value} (integer) and \verb{$prst} (XML preset name).
Alias: \code{MSO_CONNECTOR}.
}
\keyword{datasets}
