% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{Package}
\alias{Package}
\title{PowerPoint package}
\description{
Top-level container for a .pptx file. Extends OpcPackage with
presentation-specific behavior.
}
\keyword{internal}
\section{Super class}{
\code{\link[rpptx:OpcPackage]{rpptx::OpcPackage}} -> \code{Package}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Package-next_image_partname}{\code{Package$next_image_partname()}}
\item \href{#method-Package-next_media_partname}{\code{Package$next_media_partname()}}
\item \href{#method-Package-get_or_add_media_part}{\code{Package$get_or_add_media_part()}}
\item \href{#method-Package-get_or_add_image_part}{\code{Package$get_or_add_image_part()}}
\item \href{#method-Package-iter_parts}{\code{Package$iter_parts()}}
\item \href{#method-Package-clone}{\code{Package$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="drop_rel"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-drop_rel'><code>rpptx::OpcPackage$drop_rel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="initialize"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-initialize'><code>rpptx::OpcPackage$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="iter_rels"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-iter_rels'><code>rpptx::OpcPackage$iter_rels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="next_partname"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-next_partname'><code>rpptx::OpcPackage$next_partname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="part_related_by"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-part_related_by'><code>rpptx::OpcPackage$part_related_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="relate_to"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-relate_to'><code>rpptx::OpcPackage$relate_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="related_part"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-related_part'><code>rpptx::OpcPackage$related_part()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="save"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-save'><code>rpptx::OpcPackage$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpptx" data-topic="OpcPackage" data-id="target_ref"><a href='../../rpptx/html/OpcPackage.html#method-OpcPackage-target_ref'><code>rpptx::OpcPackage$target_ref()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Package-next_image_partname"></a>}}
\if{latex}{\out{\hypertarget{method-Package-next_image_partname}{}}}
\subsection{Method \code{next_image_partname()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Package$next_image_partname(ext)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Package-next_media_partname"></a>}}
\if{latex}{\out{\hypertarget{method-Package-next_media_partname}{}}}
\subsection{Method \code{next_media_partname()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Package$next_media_partname(ext)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Package-get_or_add_media_part"></a>}}
\if{latex}{\out{\hypertarget{method-Package-get_or_add_media_part}{}}}
\subsection{Method \code{get_or_add_media_part()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Package$get_or_add_media_part(video)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Package-get_or_add_image_part"></a>}}
\if{latex}{\out{\hypertarget{method-Package-get_or_add_image_part}{}}}
\subsection{Method \code{get_or_add_image_part()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Package$get_or_add_image_part(image_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Package-iter_parts"></a>}}
\if{latex}{\out{\hypertarget{method-Package-iter_parts}{}}}
\subsection{Method \code{iter_parts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Package$iter_parts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Package-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Package-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Package$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
