% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_functions.R
\name{presentation}
\alias{presentation}
\title{Create Presentation object from a PowerPoint file.}
\usage{
presentation(path)
}
\arguments{
\item{path}{(character) File path for a PowerPoint file.}
}
\value{
Python pptx Presentation object.
}
\description{
Create Presentation object from a PowerPoint file.
}
